==> 1_Compile_RSEM_out.ipynb <==
library(tidyverse)
library(stringr)
library(biomaRt)

print(sessionInfo())

setwd("./rsem_out/")

s1_3046_S1_ctrl_1_a <- read.delim("./3046_S1.genes.results", header=T, sep='\t')
s2_3046_S2_100mM_H_CPD_1_a <- read.delim("./3046_S2.genes.results", header=T, sep='\t')
s4_3046_S4_ctrl_1_b <- read.delim("./3046_S4.genes.results", header=T, sep='\t')
s5_3046_S5_100mM_H_CPD_1_b <- read.delim("./3046_S5.genes.results", header=T, sep='\t')
s7_3134_S4_ctrl_1_c <- read.delim("./3134_S4.genes.results", header=T, sep='\t')
s8_3134_S5_100mM_H_CPD_1_c <- read.delim("./3134_S5.genes.results", header=T, sep='\t')

s10_3046_S7_ctrl_2_a <- read.delim("./3046_S7.genes.results", header=T, sep='\t')
s11_3046_S8_10mM_H_MTMT_2_a <- read.delim("./3046_S8.genes.results", header=T, sep='\t')
s13_3046_S10_ctrl_2_b <- read.delim("./3046_S10.genes.results", header=T, sep='\t')
s14_3046_S11_10mM_H_MTMT_2_b <- read.delim("./3046_S11.genes.results", header=T, sep='\t')
s16_3134_S1_ctrl_2_c <- read.delim("./3134_S1.genes.results", header=T, sep='\t')
s17_3134_S2_10mM_H_MTMT_2_c <- read.delim("./3134_S2.genes.results", header=T, sep='\t')

s19_3134_S7_ctrl_3_a <- read.delim("./3134_S7.genes.results", header=T, sep='\t')
s20_3134_S8_1p_H_benzaldehyde_3_a <- read.delim("./3134_S8.genes.results", header=T, sep='\t')
s22_3134_S10_ctrl_3_b <- read.delim("./3134_S10.genes.results", header=T, sep='\t')
s23_3134_S11_1p_H_benzaldehyde_3_b <- read.delim("./3134_S11.genes.results", header=T, sep='\t')
s25_3284_S10_ctrl_3_c <- read.delim("./3284_S10.genes.results", header=T, sep='\t')
s26_3284_S11_1p_H_benzaldehyde_3_c <- read.delim("./3284_S11.genes.results", header=T, sep='\t')

s40_3552_S1_ctrl_5_a <- read.delim("./3552_S1.genes.results", header=T, sep='\t')
s41_3552_S2_1p_H_octanal_5_a <- read.delim("./3552_S2.genes.results", header=T, sep='\t')
s43_3552_S4_ctrl_5_b <- read.delim("./3552_S4.genes.results", header=T, sep='\t')
s44_3552_S5_1p_H_octanal_5_b <- read.delim("./3552_S5.genes.results", header=T, sep='\t')
s46_3606_S1_ctrl_5_c <- read.delim("./3606a_S1.genes.results", header=T, sep='\t')
s47_3606_S2_1p_H_octanal_5_c <- read.delim("./3606b_S2.genes.results", header=T, sep='\t')

s49_3552_S7_ctrl_6_a <- read.delim("./3552_S7.genes.results", header=T, sep='\t')
s50_3552_S8_1p_pDihydrocarveol_6_a <- read.delim("./3552_S8.genes.results", header=T, sep='\t')
s51_3552_S9_1p_nDihydrocarveol_6_a <- read.delim("./3552_S9.genes.results", header=T, sep='\t')
s53_3639_S1_ctrl_6_b <- read.delim("./3639_S1.genes.results", header=T, sep='\t')
s54_3639_S2_1p_pDihydrocarveol_6_b <- read.delim("./3639_S2.genes.results", header=T, sep='\t')
s55_3639_S3_1p_nDihydrocarveol_6_b <- read.delim("./3639_S3.genes.results", header=T, sep='\t')
s57_3639_S5_ctrl_6_c <- read.delim("./3639_S5.genes.results", header=T, sep='\t')
s58_3639_S6_1p_pDihydrocarveol_6_c <- read.delim("./3639_S6.genes.results", header=T, sep='\t')
s59_3639_S7_1p_nDihydrocarveol_6_c <- read.delim("./3639_S7.genes.results", header=T, sep='\t')

s61_3606_S4_ctrl_7_a <- read.delim("./3606b_S4.genes.results", header=T, sep='\t')
s62_3606_S5_1p_citronellol_7_a <- read.delim("./3606a_S5.genes.results", header=T, sep='\t')
s63_3606_S6_1_guaiacol_7_a <- read.delim("./3606b_S6.genes.results", header=T, sep='\t')
s64_3606_S7_ctrl_7_b <- read.delim("./3606b_S7.genes.results", header=T, sep='\t')
s65_3606_S8_1p_citronellol_7_b <- read.delim("./3606a_S8.genes.results", header=T, sep='\t')
s66_3606_S9_1_guaiacol_7_b <- read.delim("./3606b_S9.genes.results", header=T, sep='\t')
s67_3606_S10_ctrl_7_c <- read.delim("./3606b_S10.genes.results", header=T, sep='\t')
s68_3606_S11_1p_citronellol_7_c <- read.delim("./3606a_S11.genes.results", header=T, sep='\t')
s69_3606_S12_1_guaiacol_7_c <- read.delim("./3606a_S12.genes.results", header=T, sep='\t')

s70_3645_S1_ctrl_8_a <- read.delim("./3645_S1.genes.results", header=T, sep='\t')
s71_3645_S2_1p_pCarvone_8_a <- read.delim("./3645_S2.genes.results", header=T, sep='\t')
s72_3645_S3_1p_nCarvone_8_a <- read.delim("./3645_S3.genes.results", header=T, sep='\t')
s74_3645_S5_ctrl_8_b <- read.delim("./3645_S5.genes.results", header=T, sep='\t')
s75_3645_S6_1p_pCarvone_8_b <- read.delim("./3645_S6.genes.results", header=T, sep='\t')
s76_3645_S7_1p_nCarvone_8_b <- read.delim("./3645_S7.genes.results", header=T, sep='\t')
s78_3645_S9_ctrl_8_c <- read.delim("./3645_S9.genes.results", header=T, sep='\t')
s79_3645_S10_1p_pCarvone_8_c <- read.delim("./3645_S10.genes.results", header=T, sep='\t')
s80_3645_S11_1p_nCarvone_8_c <- read.delim("./3645_S11.genes.results", header=T, sep='\t')

s82_3769_S1_ctrl_9_a <- read.delim("./3769a_S1.genes.results", header=T, sep='\t')
s83_3769_S2_1p_isoamylAcetate_9_a <- read.delim("./3769b_S2.genes.results", header=T, sep='\t')
s84_3769_S3_p1_isovalericAcid_9_a <- read.delim("./3769a_S3.genes.results", header=T, sep='\t')
s85_3769_S4_3mM_androstenone_9_a <- read.delim("./3769a_S4.genes.results", header=T, sep='\t')
s86_3769_S5_ctrl_9_b <- read.delim("./3769a_S5.genes.results", header=T, sep='\t')
s87_3769_S6_1p_isoamylAcetate_9_b <- read.delim("./3769a_S6.genes.results", header=T, sep='\t')
s88_3769_S7_p1_isovalericAcid_9_b <- read.delim("./3769a_S7.genes.results", header=T, sep='\t')
s89_3769_S8_3mM_androstenone_9_b <- read.delim("./3769a_S8.genes.results", header=T, sep='\t')
s90_3769_S9_ctrl_9_c <- read.delim("./3769b_S9.genes.results", header=T, sep='\t')
s91_3769_S10_1p_isoamylAcetate_9_c <- read.delim("./3769b_S10.genes.results", header=T, sep='\t')
s92_3769_S11_p1_isovalericAcid_9_c <- read.delim("./3769a_S11.genes.results", header=T, sep='\t')
s93_3769_S12_3mM_androstenone_9_c <- read.delim("./3769a_S12.genes.results", header=T, sep='\t')

s94_4258_S1_ctrl_10_a <- read.delim("./4258_S1.genes.results", header=T, sep='\t')
s96_4258_S3_1p_bCaryophyllene_10_a <- read.delim("./4258_S3.genes.results", header=T, sep='\t')
s97_4258_S4_1p_ethylTiglate_10_a <- read.delim("./4258_S4.genes.results", header=T, sep='\t')

s98_4258_S5_ctrl_10_b <- read.delim("./4258_S5.genes.results", header=T, sep='\t')
s99_4258_S6_1p_pyridine_10_b <- read.delim("./4258_S6.genes.results", header=T, sep='\t')
s100_4258_S7_1p_bCaryophyllene_10_b <- read.delim("./4258_S7.genes.results", header=T, sep='\t')
s101_4258_S8_1p_methylSalicylate_10_b <- read.delim("./4258_S8.genes.results", header=T, sep='\t')

s102_4258_S9_ctrl_10_c <- read.delim("./4258_S9.genes.results", header=T, sep='\t')
s103_4258_S10_1p_pyridine_10_c <- read.delim("./4258_S10.genes.results", header=T, sep='\t')
s104_4258_S11_1p_bCaryophyllene_10_c <- read.delim("./4258_S11.genes.results", header=T, sep='\t')

s106_4259_S1_ctrl_11_a <- read.delim("./4259_S1.genes.results", header=T, sep='\t')
s107_4259_S2_1p_2heptanone_11_a <- read.delim("./4259_S2.genes.results", header=T, sep='\t')
s108_4259_S3_1p_linalool_11_a <- read.delim("./4259_S3.genes.results", header=T, sep='\t')
s109_4259_S4_1p_ethylButyrate_11_a <- read.delim("./4259_S4.genes.results", header=T, sep='\t')

s110_4259_S5_ctrl_11_b <- read.delim("./4259_S5.genes.results", header=T, sep='\t')
s111_4259_S6_1p_2heptanone_11_b <- read.delim("./4259_S6.genes.results", header=T, sep='\t')
s112_4259_S7_1p_linalool_11_b <- read.delim("./4259_S7.genes.results", header=T, sep='\t')
s113_4259_S8_1p_ethylButyrate_11_b <- read.delim("./4259_S8.genes.results", header=T, sep='\t')

s114_4259_S9_ctrl_11_c <- read.delim("./4259_S9.genes.results", header=T, sep='\t')
s116_4259_S11_1p_linalool_11_c <- read.delim("./4259_S11.genes.results", header=T, sep='\t')
s117_4259_S12_1p_ethylButyrate_11_c <- read.delim("./4259_S12.genes.results", header=T, sep='\t')

s118_4269_S1_ctrl_12_a <- read.delim("./4269_S1.genes.results", header=T, sep='\t')
s119_4269_S2_1M_pMenthol_12_a <- read.delim("./4269_S2.genes.results", header=T, sep='\t')
s120_4269_S3_1M_nMenthol_12_a <- read.delim("./4269_S3.genes.results", header=T, sep='\t')
s121_4269_S4_ctrl_12_b <- read.delim("./4269_S4.genes.results", header=T, sep='\t')
s122_4269_S5_1M_pMenthol_12_b <- read.delim("./4269_S5.genes.results", header=T, sep='\t')
s123_4269_S6_1M_nMenthol_12_b <- read.delim("./4269_S6.genes.results", header=T, sep='\t')
s124_4269_S7_ctrl_12_c <- read.delim("./4269_S7.genes.results", header=T, sep='\t')
s125_4269_S8_1M_pMenthol_12_c <- read.delim("./4269_S8.genes.results", header=T, sep='\t')
s126_4269_S9_1M_nMenthol_12_c <- read.delim("./4269_S9.genes.results", header=T, sep='\t')

s127_4269_S10_ctrl_13_c <- read.delim("./4269_S10.genes.results", header=T, sep='\t')
s128_4269_S11_1p_2heptanone_13_c <- read.delim("./4269_S11.genes.results", header=T, sep='\t')
s129_4269_S12_1p_pyridine_13_c <- read.delim("./4269_S12.genes.results", header=T, sep='\t')

s130_4331_S1_ctrl_14_a <- read.delim("./4331_S1.genes.results", header=T, sep='\t')
s131_4331_S2_100p_ebFarnesene_14_a <- read.delim("./4331_S2.genes.results", header=T, sep='\t')
s132_4331_S3_1p_25dmp_14_a <- read.delim("./4331_S3.genes.results", header=T, sep='\t')
s133_4331_S4_100uM_MTMT_14_a <- read.delim("./4331_S4.genes.results", header=T, sep='\t')
s134_4331_S5_ctrl_14_b <- read.delim("./4331_S5.genes.results", header=T, sep='\t')
s135_4331_S6_100p_ebFarnesene_14_b <- read.delim("./4331_S6.genes.results", header=T, sep='\t')
s136_4331_S7_1p_25dmp_14_b <- read.delim("./4331_S7.genes.results", header=T, sep='\t')
s137_4331_S8_100uM_MTMT_14_b <- read.delim("./4331_S8.genes.results", header=T, sep='\t')
s138_4331_S9_ctrl_14_c <- read.delim("./4331_S9.genes.results", header=T, sep='\t')
s139_4331_S10_100p_ebFarnesene_14_c <- read.delim("./4331_S10.genes.results", header=T, sep='\t')
s140_4331_S11_1p_25dmp_14_c <- read.delim("./4331_S11.genes.results", header=T, sep='\t')
s141_4331_S12_100uM_MTMT_14_c <- read.delim("./4331_S12.genes.results", header=T, sep='\t')

s142_4429_S5_ctrl_15_a <- read.delim("./4429_S5.genes.results", header=T, sep='\t')
s143_4429_S6_10p_p2octanol_15_a <- read.delim("./4429_S6.genes.results", header=T, sep='\t')
s144_4429_S7_10p_n2octanol_15_a <- read.delim("./4429_S7.genes.results", header=T, sep='\t')
s145_4429_S8_77p_dhb_15_a <- read.delim("./4429_S8.genes.results", header=T, sep='\t')
s146_4429_S9_ctrl_15_b <- read.delim("./4429_S9.genes.results", header=T, sep='\t')
s147_4429_S10_10p_p2octanol_15_b <- read.delim("./4429_S10.genes.results", header=T, sep='\t')
s148_4429_S11_10p_n2octanol_15_b <- read.delim("./4429_S11.genes.results", header=T, sep='\t')
s149_4429_S12_77p_dhb_15_b <- read.delim("./4429_S12.genes.results", header=T, sep='\t')
s150_4575_S1_ctrl_15_c <- read.delim("./4575_S1.genes.results", header=T, sep='\t')
s151_4998_S7_10p_p2octanol_15_c <- read.delim("./4998_S7.genes.results", header=T, sep='\t')
s152_4998_S8_10p_n2octanol_15_c <- read.delim("./4998_S8.genes.results", header=T, sep='\t')
s153_4575_S2_77p_dhb_15_c <- read.delim("./4575_S2.genes.results", header=T, sep='\t')

s158_4761_S1_ctrl_17_a <- read.delim("./4761_S1.genes.results", header=T, sep='\t')
s159_4761_S2_p01_sbt_17_a <- read.delim("./4761_S2.genes.results", header=T, sep='\t')
s160_4761_S3_1p_aPinene_17_a <- read.delim("./4761_S3.genes.results", header=T, sep='\t')
s161_4761_S4_1p_bDamascone_17_a <- read.delim("./4761_S4.genes.results", header=T, sep='\t')
s162_4761_S5_ctrl_17_b <- read.delim("./4761_S5.genes.results", header=T, sep='\t')
s163_4761_S6_p01_sbt_17_b <- read.delim("./4761_S6.genes.results", header=T, sep='\t')
s164_4761_S7_1p_aPinene_17_b <- read.delim("./4761_S7.genes.results", header=T, sep='\t')
s165_4761_S8_1p_bDamascone_17_b <- read.delim("./4761_S8.genes.results", header=T, sep='\t')
s166_4761_S9_ctrl_17_c <- read.delim("./4761_S9.genes.results", header=T, sep='\t')
s167_4761_S10_p01_sbt_17_c <- read.delim("./4761_S10.genes.results", header=T, sep='\t')
s168_4761_S11_1p_aPinene_17_c <- read.delim("./4761_S11.genes.results", header=T, sep='\t')
s169_4761_S12_1p_bDamascone_17_c <- read.delim("./4761_S12.genes.results", header=T, sep='\t')

s170_4999_S1_ctrl_18_a <- read.delim("./4999_S1.genes.results", header=T, sep='\t')
s171_4999_S2_1p_2m2p_18_a <- read.delim("./4999_S2.genes.results", header=T, sep='\t')
s173_4999_S4_1p_transCinnamaldehyde_18_a <- read.delim("./4999_S4.genes.results", header=T, sep='\t')
s174_4999_S5_ctrl_18_b <- read.delim("./4999_S5.genes.results", header=T, sep='\t')
s175_4999_S6_1p_2m2p_18_b <- read.delim("./4999_S6.genes.results", header=T, sep='\t')
s177_4999_S8_1p_transCinnamaldehyde_18_b <- read.delim("./4999_S8.genes.results", header=T, sep='\t')
s178_4999_S9_ctrl_18_c <- read.delim("./4999_S9.genes.results", header=T, sep='\t')
s179_4999_S10_1p_2m2p_18_c <- read.delim("./4999_S10.genes.results", header=T, sep='\t')
s181_4999_S12_1p_transCinnamaldehyde_18_c <- read.delim("./4999_S12.genes.results", header=T, sep='\t')

s182_5000_S1_ctrl_19_a <- read.delim("./5000_S1.genes.results", header=T, sep='\t')
s183_5000_S2_1p_hexylTiglate_19_a <- read.delim("./5000_S2.genes.results", header=T, sep='\t')
s184_5000_S3_1p_2ethyl3methylPyrazine_19_a <- read.delim("./5000_S3.genes.results", header=T, sep='\t')
s185_5000_S4_1p_isopropylTiglate_19_a <- read.delim("./5000_S4.genes.results", header=T, sep='\t')
s186_5000_S5_ctrl_19_b <- read.delim("./5000_S5.genes.results", header=T, sep='\t')
s187_5000_S6_1p_hexylTiglate_19_b <- read.delim("./5000_S6.genes.results", header=T, sep='\t')
s188_5000_S7_1p_2ethyl3methylPyrazine_19_b <- read.delim("./5000_S7.genes.results", header=T, sep='\t')
s189_5000_S8_1p_isopropylTiglate_19_b <- read.delim("./5000_S8.genes.results", header=T, sep='\t')
s190_5000_S9_ctrl_19_c <- read.delim("./5000_S9.genes.results", header=T, sep='\t')
s191_5000_S10_1p_hexylTiglate_19_c <- read.delim("./5000_S10.genes.results", header=T, sep='\t')
s192_5000_S11_1p_2ethyl3methylPyrazine_19_c <- read.delim("./5000_S11.genes.results", header=T, sep='\t')
s193_5000_S12_1p_isopropylTiglate_19_c <- read.delim("./5000_S12.genes.results", header=T, sep='\t')

s206_1_5mn1_ctrl_21_a <- read.delim("./1-5mn1.genes.results", header=T, sep='\t')
s207_1_5mn2_1p_2hac_21_a <- read.delim("./1-5mn2.genes.results", header=T, sep='\t')
s208_1_5mn3_1p_heptanal_21_a <- read.delim("./1-5mn3.genes.results", header=T, sep='\t')
s209_1_5mn4_ctrl_21_b <- read.delim("./1-5mn4.genes.results", header=T, sep='\t')
s210_1_5mn5_1p_2hac_21_b <- read.delim("./1-5mn5.genes.results", header=T, sep='\t')
s211_1_5mn6_1p_heptanal_21_b <- read.delim("./1-5mn6.genes.results", header=T, sep='\t')
s212_6_9mn4_ctrl_21_c <- read.delim("./6-9mn4.genes.results", header=T, sep='\t')
s213_6_9mn5_1p_2hac_21_c <- read.delim("./6-9mn5.genes.results", header=T, sep='\t')

s215_1_5mn7_ctrl_22_a <- read.delim("./1-5mn7.genes.results", header=T, sep='\t')
s216_1_5mn8_1p_4methylAcetophenone_22_a <- read.delim("./1-5mn8.genes.results", header=T, sep='\t')
s217_1_5mn9_1p_3methyl1butanethiol_22_a <- read.delim("./1-5mn9.genes.results", header=T, sep='\t')
s218_1_5mn10_ctrl_22_b <- read.delim("./1-5mn10.genes.results", header=T, sep='\t')
s219_1_5mn11_1p_4methylAcetophenone_22_b <- read.delim("./1-5mn11.genes.results", header=T, sep='\t')
s220_1_5mn12_1p_3methyl1butanethiol_22_b <- read.delim("./1-5mn12.genes.results", header=T, sep='\t')
s221_6_9mn1_ctrl_22_c <- read.delim("./6-9mn1.genes.results", header=T, sep='\t')
s222_6_9mn2_1p_4methylAcetophenone_22_c <- read.delim("./6-9mn2.genes.results", header=T, sep='\t')

s224_14_17mn1_ctrl_21a22_c <- read.delim("./14-17mn1.genes.results", header=T, sep='\t')
s225_14_17mn2_1p_3methyl1butanethiol_22_c <- read.delim("./14-17mn2.genes.results", header=T, sep='\t')
s226_14_17mn3_1p_heptanal_21_c <- read.delim("./14-17mn3.genes.results", header=T, sep='\t')

s227_6_9mn7_ctrl_23_a <- read.delim("./6-9mn7.genes.results", header=T, sep='\t')
s228_6_9mn8_1p_paraCresol_23_a <- read.delim("./6-9mn8.genes.results", header=T, sep='\t')
s229_6_9mn9_1p_dimethyltrisulfide_23_a <- read.delim("./6-9mn9.genes.results", header=T, sep='\t')
s230_6_9mn10_ctrl_23_b <- read.delim("./6-9mn10.genes.results", header=T, sep='\t')
s231_6_9mn11_1p_paraCresol_23_b <- read.delim("./6-9mn11.genes.results", header=T, sep='\t')
s232_6_9mn12_1p_dimethyltrisulfide_23_b <- read.delim("./6-9mn12.genes.results", header=T, sep='\t')
s233_10_13mn1_ctrl_23_c <- read.delim("./10-13mn1.genes.results", header=T, sep='\t')
s234_10_13mn2_1p_paraCresol_23_c <- read.delim("./10-13mn2.genes.results", header=T, sep='\t')
s235_10_13mn3_1p_dimethyltrisulfide_23_c <- read.delim("./10-13mn3.genes.results", header=T, sep='\t')

s236_10_13mn4_ctrl_24_a <- read.delim("./10-13mn4.genes.results", header=T, sep='\t')
s237_10_13mn5_1p_butyricAcid_24_a <- read.delim("./10-13mn5.genes.results", header=T, sep='\t')
s238_10_13mn6_1p_indole_24_a <- read.delim("./10-13mn6.genes.results", header=T, sep='\t')
s239_10_13mn7_ctrl_24_b <- read.delim("./10-13mn7.genes.results", header=T, sep='\t')
s240_10_13mn8_1p_butyricAcid_24_b <- read.delim("./10-13mn8.genes.results", header=T, sep='\t')
s241_10_13mn9_1p_indole_24_b <- read.delim("./10-13mn9.genes.results", header=T, sep='\t')
s242_10_13mn10_ctrl_24_c <- read.delim("./10-13mn10.genes.results", header=T, sep='\t')
s243_10_13mn11_1p_butyricAcid_24_c <- read.delim("./10-13mn11.genes.results", header=T, sep='\t')
s244_10_13mn12_1p_indole_24_c <- read.delim("./10-13mn12.genes.results", header=T, sep='\t')

s245_14_17mn4_ctrl_25_a <- read.delim("./14-17mn4.genes.results", header=T, sep='\t')
s246_14_17mn5_1p_tbm_25_a <- read.delim("./14-17mn5.genes.results", header=T, sep='\t')
s247_14_17mn6_1p_diacetyl_25_a <- read.delim("./14-17mn6.genes.results", header=T, sep='\t')
s248_14_17mn7_ctrl_25_b <- read.delim("./14-17mn7.genes.results", header=T, sep='\t')
s249_14_17mn8_1p_tbm_25_b <- read.delim("./14-17mn8.genes.results", header=T, sep='\t')
s250_14_17mn9_1p_diacetyl_25_b <- read.delim("./14-17mn9.genes.results", header=T, sep='\t')
s251_14_17mn10_ctrl_25_c <- read.delim("./14-17mn10.genes.results", header=T, sep='\t')
s252_14_17mn11_1p_tbm_25_c <- read.delim("./14-17mn11.genes.results", header=T, sep='\t')
s253_14_17mn12_1p_diacetyl_25_c <- read.delim("./14-17mn12.genes.results", header=T, sep='\t')

s254_18_1_No_ctrl_26_a <- read.delim("./18-1-No.genes.results", header=T, sep='\t')
s255_18_2_He_1p_2hexanone_26_a <- read.delim("./18-2-He.genes.results", header=T, sep='\t')
s256_18_3_PEA_1p_2phenylethylalcohol_26_a <- read.delim("./18-3-PEA.genes.results", header=T, sep='\t')
s257_19_1_No_ctrl_26_b <- read.delim("./19-1-No.genes.results", header=T, sep='\t')
s258_19_2_He_1p_2hexanone_26_b <- read.delim("./19-2-He.genes.results", header=T, sep='\t')
s259_19_3_PEA_1p_2phenylethylalcohol_26_b <- read.delim("./19-3-PEA.genes.results", header=T, sep='\t')
s260_mn1_ctrl_26_c <- read.delim("./mn1.genes.results", header=T, sep='\t')
s261_mn2_1p_2hexanone_26_c <- read.delim("./mn2.genes.results", header=T, sep='\t')
s262_mn3_1p_2phenylethylalcohol_26_c <- read.delim("./mn3.genes.results", header=T, sep='\t')

s263_mn4_ctrl_27_a <- read.delim("./mn4.genes.results", header=T, sep='\t')
s264_mn5_1p_tbm_27_a <- read.delim("./mn5.genes.results", header=T, sep='\t')
s265_mn6_1p_acetophenone_27_a <- read.delim("./mn6.genes.results", header=T, sep='\t')
s266_mn7_ctrl_27_b <- read.delim("./mn7.genes.results", header=T, sep='\t')
s267_mn8_1p_tbm_27_b <- read.delim("./mn8.genes.results", header=T, sep='\t')
s268_mn9_1p_acetophenone_27_b <- read.delim("./mn9.genes.results", header=T, sep='\t')
s269_mn10_ctrl_27_c <- read.delim("./mn10.genes.results", header=T, sep='\t')
s270_mn11_1p_tbm_27_c <- read.delim("./mn11.genes.results", header=T, sep='\t')
s271_mn12_1p_acetophenone_27_c <- read.delim("./mn12.genes.results", header=T, sep='\t')

s272_24_1_No_ctrl_28_a <- read.delim("./24-1-No.genes.results", header=T, sep='\t')
s273_24_2_Sk_p01_e2butene1thiol_28_a <- read.delim("./24-2-Sk.genes.results", header=T, sep='\t')
s274_24_3_Cy_p01_cyclopentanethiol_28_a <- read.delim("./24-3-Cy.genes.results", header=T, sep='\t')
s275_25_1_No_ctrl_28_b <- read.delim("./25-1-No.genes.results", header=T, sep='\t')
s276_25_2_Sk_p01_e2butene1thiol_28_b <- read.delim("./25-2-Sk.genes.results", header=T, sep='\t')
s277_25_3_Cy_p01_cyclopentanethiol_28_b <- read.delim("./25-3-Cy.genes.results", header=T, sep='\t')
s278_26_1_No_ctrl_28_c <- read.delim("./26-1-No.genes.results", header=T, sep='\t')
s279_26_2_Sk_p01_e2butene1thiol_28_c <- read.delim("./26-2-Sk.genes.results", header=T, sep='\t')
s280_26_3_Cy_p01_cyclopentanethiol_28_c <- read.delim("./26-3-Cy.genes.results", header=T, sep='\t')

s281_27_1_No_ctrl_29_a <- read.delim("./27-1-No.genes.results", header=T, sep='\t')
s282_27_2_Ci_1p_citral_29_a <- read.delim("./27-2-Ci.genes.results", header=T, sep='\t')
s283_27_3_An_1p_anisaldehyde_29_a <- read.delim("./27-3-An.genes.results", header=T, sep='\t')
s284_28_1_No_ctrl_29_b <- read.delim("./28-1-No.genes.results", header=T, sep='\t')
s285_28_2_Ci_1p_citral_29_b <- read.delim("./28-2-Ci.genes.results", header=T, sep='\t')
s286_28_3_An_1p_anisaldehyde_29_b <- read.delim("./28-3-An.genes.results", header=T, sep='\t')
s287_29_1_No_ctrl_29_c <- read.delim("./29-1-No.genes.results", header=T, sep='\t')
s288_29_2_Ci_1p_citral_29_c <- read.delim("./29-2-Ci.genes.results", header=T, sep='\t')
s289_29_3_An_1p_anisaldehyde_29_c <- read.delim("./29-3-An.genes.results", header=T, sep='\t')

s290_SerA4_TCCTGAGC_ctrl_30_a <- read.delim("./SerA4_TCCTGAGC.genes.results", header=T, sep='\t')
s291_SerA5_GGACTCCT_1p_acetophenone_30_a <- read.delim("./SerA5_GGACTCCT.genes.results", header=T, sep='\t')
s292_SerA6_TAGGCATG_100p_acetophenone_30_a <- read.delim("./SerA6_TAGGCATG.genes.results", header=T, sep='\t')
s293_SerB4_TCCTGAGC_ctrl_30_b <- read.delim("./SerB4_TCCTGAGC.genes.results", header=T, sep='\t')
s294_SerB5_GGACTCCT_1p_acetophenone_30_b <- read.delim("./SerB5_GGACTCCT.genes.results", header=T, sep='\t')
s295_SerB6_TAGGCATG_100p_acetophenone_30_b <- read.delim("./SerB6_TAGGCATG.genes.results", header=T, sep='\t')
s296_SerB10_CGAGGCTG_ctrl_30_c <- read.delim("./SerB10_CGAGGCTG.genes.results", header=T, sep='\t')
s297_SerB11_AAGAGGCA_1p_acetophenone_30_c <- read.delim("./SerB11_AAGAGGCA.genes.results", header=T, sep='\t')
s298_SerB12_GTAGAGGA_100p_acetophenone_30_c <- read.delim("./SerB12_GTAGAGGA.genes.results", header=T, sep='\t')

s299_SerA7_CTCTCTAC_ctrl_31_a <- read.delim("./SerA7_CTCTCTAC.genes.results", header=T, sep='\t')
s300_SerA8_CAGAGAGG_1p_tmt_31_a <- read.delim("./SerA8_CAGAGAGG.genes.results", header=T, sep='\t')
s301_SerA9_GCTACGCT_100p_tmt_31_a <- read.delim("./SerA9_GCTACGCT.genes.results", header=T, sep='\t')
s302_SerA10_CGAGGCTG_ctrl_31_b <- read.delim("./SerA10_CGAGGCTG.genes.results", header=T, sep='\t')
s303_SerA11_AAGAGGCA_1p_tmt_31_b <- read.delim("./SerA11_AAGAGGCA.genes.results", header=T, sep='\t')
s304_SerA12_GTAGAGGA_100p_tmt_31_b <- read.delim("./SerA12_GTAGAGGA.genes.results", header=T, sep='\t')
s305_SerB1_TAAGGCGA_ctrl_31_c <- read.delim("./SerB1_TAAGGCGA.genes.results", header=T, sep='\t')
s306_SerB2_CGTACTAG_1p_tmt_31_c <- read.delim("./SerB2_CGTACTAG.genes.results", header=T, sep='\t')
s307_SerB3_AGGCAGAA_100p_tmt_31_c <- read.delim("./SerB3_AGGCAGAA.genes.results", header=T, sep='\t')

s308_SerB7_CTCTCTAC_ctrl_32_a <- read.delim("./SerB7_CTCTCTAC.genes.results", header=T, sep='\t')
s309_SerB8_CAGAGAGG_1p_2m2t_32_a <- read.delim("./SerB8_CAGAGAGG.genes.results", header=T, sep='\t')
s310_SerB9_GCTACGCT_100p_2m2t_32_a <- read.delim("./SerB9_GCTACGCT.genes.results", header=T, sep='\t')
s311_SerC1_TAAGGCGA_ctrl_32_b <- read.delim("./SerC1_TAAGGCGA.genes.results", header=T, sep='\t')
s312_SerC2_CGTACTAG_1p_2m2t_32_b <- read.delim("./SerC2_CGTACTAG.genes.results", header=T, sep='\t')
s313_SerC3_AGGCAGAA_100p_2m2t_32_b <- read.delim("./SerC3_AGGCAGAA.genes.results", header=T, sep='\t')
s314_SerC4_TCCTGAGC_ctrl_32_c <- read.delim("./SerC4_TCCTGAGC.genes.results", header=T, sep='\t')
s315_SerC5_GGACTCCT_1p_2m2t_32_c <- read.delim("./SerC5_GGACTCCT.genes.results", header=T, sep='\t')
s316_SerC6_TAGGCATG_100p_2m2t_32_c <- read.delim("./SerC6_TAGGCATG.genes.results", header=T, sep='\t')

s317_SerC7_CTCTCTAC_ctrl_33_a <- read.delim("./SerC7_CTCTCTAC.genes.results", header=T, sep='\t')
s318_SerC8_CAGAGAGG_1p_sbt_33_a <- read.delim("./SerC8_CAGAGAGG.genes.results", header=T, sep='\t')
s319_SerC9_GCTACGCT_100p_sbt_33_a <- read.delim("./SerC9_GCTACGCT.genes.results", header=T, sep='\t')
s320_SerC10_CGAGGCTG_ctrl_33_b <- read.delim("./SerC10_CGAGGCTG.genes.results", header=T, sep='\t')
s321_SerC11_AAGAGGCA_1p_sbt_33_b <- read.delim("./SerC11_AAGAGGCA.genes.results", header=T, sep='\t')
s322_SerC12_GTAGAGGA_100p_sbt_33_b <- read.delim("./SerC12_GTAGAGGA.genes.results", header=T, sep='\t')
s323_SerD01_ctrl_33_c <- read.delim("./SerD01.genes.results", header=T, sep='\t')
s324_SerD02_1p_sbt_33_c <- read.delim("./SerD02.genes.results", header=T, sep='\t')
s325_SerD03_100p_sbt_33_c <- read.delim("./SerD03.genes.results", header=T, sep='\t')

s326_SerD04_ctrl_34_a <- read.delim("./SerD04.genes.results", header=T, sep='\t')
s327_SerD05_1p_ntmt_34_a <- read.delim("./SerD05.genes.results", header=T, sep='\t')
s329_SerD07_ctrl_34_b <- read.delim("./SerD07.genes.results", header=T, sep='\t')
s330_SerD08_1p_ntmt_34_b <- read.delim("./SerD08.genes.results", header=T, sep='\t')
s332_SerD10_ctrl_34_c <- read.delim("./SerD10.genes.results", header=T, sep='\t')
s333_SerD11_1p_ntmt_34_c <- read.delim("./SerD11.genes.results", header=T, sep='\t')

s335_SerE1_L001_R1_001_ctrl_35_a <- read.delim("./SerE1_L001_R1_001.genes.results", header=T, sep='\t')
s336_SerE2_L001_R1_001_p01_acetophenone_35_a <- read.delim("./SerE2_L001_R1_001.genes.results", header=T, sep='\t')
s337_SerE3_L001_R1_001_p1_acetophenone_35_a <- read.delim("./SerE3_L001_R1_001.genes.results", header=T, sep='\t')
s338_SerE4_L001_R1_001_ctrl_35_b <- read.delim("./SerE4_L001_R1_001.genes.results", header=T, sep='\t')
s339_SerE5_L001_R1_001_p01_acetophenone_35_b <- read.delim("./SerE5_L001_R1_001.genes.results", header=T, sep='\t')
s340_SerE6_L001_R1_001_p1_acetophenone_35_b <- read.delim("./SerE6_L001_R1_001.genes.results", header=T, sep='\t')
s341_SerE7_L001_R1_001_ctrl_35_c <- read.delim("./SerE7_L001_R1_001.genes.results", header=T, sep='\t')
s342_SerE8_L001_R1_001_p01_acetophenone_35_c <- read.delim("./SerE8_L001_R1_001.genes.results", header=T, sep='\t')
s343_SerE9_L001_R1_001_p1_acetophenone_35_c <- read.delim("./SerE9_L001_R1_001.genes.results", header=T, sep='\t')

s344_SerE10_L001_R1_001_ctrl_10_a <- read.delim("./SerE10_L001_R1_001.genes.results", header=T, sep='\t')
s345_SerE11_L001_R1_001_1p_methylSalicylate_10_a <- read.delim("./SerE11_L001_R1_001.genes.results", header=T, sep='\t')
s347_SerH10_combine_ctrl_10_b <- read.delim("./SerH10_combine.genes.results", header=T, sep='\t')
s348_SerH11_combine_1p_methylSalicylate_10_b <- read.delim("./SerH11_combine.genes.results", header=T, sep='\t')

s350_SerG1_L001_R1_001_ctrl_36_a <- read.delim("./SerG1_L001_R1_001.genes.results", header=T, sep='\t')
s351_SerG2_L001_R1_001_p01_tmt_36_a <- read.delim("./SerG2_L001_R1_001.genes.results", header=T, sep='\t')
s352_SerG3_L001_R1_001_p1_tmt_36_a <- read.delim("./SerG3_L001_R1_001.genes.results", header=T, sep='\t')
s353_SerG4_L001_R1_001_10p_tmt_36_a <- read.delim("./SerG4_L001_R1_001.genes.results", header=T, sep='\t')
s354_SerG5_L001_R1_001_ctrl_36_b <- read.delim("./SerG5_L001_R1_001.genes.results", header=T, sep='\t')
s355_SerG6_L001_R1_001_p01_tmt_36_b <- read.delim("./SerG6_L001_R1_001.genes.results", header=T, sep='\t')
s356_SerG7_L001_R1_001_p1_tmt_36_b <- read.delim("./SerG7_L001_R1_001.genes.results", header=T, sep='\t')
s357_SerG8_L001_R1_001_10p_tmt_36_b <- read.delim("./SerG8_L001_R1_001.genes.results", header=T, sep='\t')
s358_SerG9_L001_R1_001_ctrl_36_c <- read.delim("./SerG9_L001_R1_001.genes.results", header=T, sep='\t')
s359_SerG10_L001_R1_001_p01_tmt_36_c <- read.delim("./SerG10_L001_R1_001.genes.results", header=T, sep='\t')
s360_SerG11_L001_R1_001_p1_tmt_36_c <- read.delim("./SerG11_L001_R1_001.genes.results", header=T, sep='\t')
s361_SerG12_L001_R1_001_10p_tmt_36_c <- read.delim("./SerG12_L001_R1_001.genes.results", header=T, sep='\t')

s362_SerH01_combine_ctrl_37_a <- read.delim("./SerH01_combine.genes.results", header=T, sep='\t')
s363_SerH02_combine_1p_pLimonene_37_a <- read.delim("./SerH02_combine.genes.results", header=T, sep='\t')
s364_SerH03_combine_1p_propionicAcid_37_a <- read.delim("./SerH03_combine.genes.results", header=T, sep='\t')
s365_SerH04_combine_ctrl_37_b <- read.delim("./SerH04_combine.genes.results", header=T, sep='\t')
s366_SerH05_combine_1p_pLimonene_37_b <- read.delim("./SerH05_combine.genes.results", header=T, sep='\t')
s367_SerH06_combine_1p_propionicAcid_37_b <- read.delim("./SerH06_combine.genes.results", header=T, sep='\t')
s368_SerH07_combine_ctrl_37_c <- read.delim("./SerH07_combine.genes.results", header=T, sep='\t')
s369_SerH08_combine_1p_pLimonene_37_c <- read.delim("./SerH08_combine.genes.results", header=T, sep='\t')
s370_SerH09_combine_1p_propionicAcid_37_c <- read.delim("./SerH09_combine.genes.results", header=T, sep='\t')

s371_SerJ5_12072017_ctrl_38_a <- read.delim("./SerJ5_12072017.genes.results", header=T, sep='\t')
s372_SerJ6_12072017_1p_geraniol_38_a <- read.delim("./SerJ6_12072017.genes.results", header=T, sep='\t')
s373_SerJ7_12072017_1p_bIonone_38_a <- read.delim("./SerJ7_12072017.genes.results", header=T, sep='\t')
s374_SerJ8_12072017_1p_dimethylsulfide_38_a <- read.delim("./SerJ8_12072017.genes.results", header=T, sep='\t')
s375_SerJ9_12072017_ctrl_38_b <- read.delim("./SerJ9_12072017.genes.results", header=T, sep='\t')
s376_SerJ10_12072017_1p_geraniol_38_b <- read.delim("./SerJ10_12072017.genes.results", header=T, sep='\t')
s377_SerJ11_12072017_1p_bIonone_38_b <- read.delim("./SerJ11_12072017.genes.results", header=T, sep='\t')
s378_SerJ12_12072017_1p_dimethylsulfide_38_b <- read.delim("./SerJ12_12072017.genes.results", header=T, sep='\t')
s379_SerF01_combine_ctrl_38_c <- read.delim("./SerF01_combine.genes.results", header=T, sep='\t')
s380_SerF02_combine_1p_geraniol_38_c <- read.delim("./SerF02_combine.genes.results", header=T, sep='\t')
s381_SerF03_combine_1p_bIonone_38_c <- read.delim("./SerF03_combine.genes.results", header=T, sep='\t')
s382_SerF04_combine_1p_dimethylsulfide_38_c <- read.delim("./SerF04_combine.genes.results", header=T, sep='\t')

s383_SerF05_combine_ctrl_39_a <- read.delim("./SerF05_combine.genes.results", header=T, sep='\t')
s384_SerF06_combine_1p_ethylTiglate_39_a <- read.delim("./SerF06_combine.genes.results", header=T, sep='\t')
s386_SerF08_combine_1p_heptanoicAcid_39_a <- read.delim("./SerF08_combine.genes.results", header=T, sep='\t')
s387_SerF09_combine_ctrl_39_b <- read.delim("./SerF09_combine.genes.results", header=T, sep='\t')
s388_SerF10_combine_1p_ethylTiglate_39_b <- read.delim("./SerF10_combine.genes.results", header=T, sep='\t')
s390_SerF12_combine_1p_heptanoicAcid_39_b <- read.delim("./SerF12_combine.genes.results", header=T, sep='\t')
s391_SerJ1_12072017_ctrl_39_c <- read.delim("./SerJ1_12072017.genes.results", header=T, sep='\t')
s392_SerJ2_12072017_1p_ethylTiglate_39_c <- read.delim("./SerJ2_12072017.genes.results", header=T, sep='\t')
s394_SerJ4_12072017_1p_heptanoicAcid_39_c <- read.delim("./SerJ4_12072017.genes.results", header=T, sep='\t')


compiledDataCounts <- data.frame(
    s1_3046_S1_ctrl_1_a = s1_3046_S1_ctrl_1_a$expected_count,
    s2_3046_S2_100mM_H_CPD_1_a = s2_3046_S2_100mM_H_CPD_1_a$expected_count,
    s4_3046_S4_ctrl_1_b = s4_3046_S4_ctrl_1_b$expected_count,
    s5_3046_S5_100mM_H_CPD_1_b = s5_3046_S5_100mM_H_CPD_1_b$expected_count,
    s7_3134_S4_ctrl_1_c = s7_3134_S4_ctrl_1_c$expected_count,
    s8_3134_S5_100mM_H_CPD_1_c = s8_3134_S5_100mM_H_CPD_1_c$expected_count,
    s10_3046_S7_ctrl_2_a = s10_3046_S7_ctrl_2_a$expected_count,
    s11_3046_S8_10mM_H_MTMT_2_a = s11_3046_S8_10mM_H_MTMT_2_a$expected_count,
    s13_3046_S10_ctrl_2_b = s13_3046_S10_ctrl_2_b$expected_count,
    s14_3046_S11_10mM_H_MTMT_2_b = s14_3046_S11_10mM_H_MTMT_2_b$expected_count,
    s16_3134_S1_ctrl_2_c = s16_3134_S1_ctrl_2_c$expected_count,
    s17_3134_S2_10mM_H_MTMT_2_c = s17_3134_S2_10mM_H_MTMT_2_c$expected_count,
    s19_3134_S7_ctrl_3_a = s19_3134_S7_ctrl_3_a$expected_count,
    s20_3134_S8_1p_H_benzaldehyde_3_a = s20_3134_S8_1p_H_benzaldehyde_3_a$expected_count,
    s22_3134_S10_ctrl_3_b = s22_3134_S10_ctrl_3_b$expected_count,
    s23_3134_S11_1p_H_benzaldehyde_3_b = s23_3134_S11_1p_H_benzaldehyde_3_b$expected_count,
    s25_3284_S10_ctrl_3_c = s25_3284_S10_ctrl_3_c$expected_count,
    s26_3284_S11_1p_H_benzaldehyde_3_c = s26_3284_S11_1p_H_benzaldehyde_3_c$expected_count,
    s40_3552_S1_ctrl_5_a = s40_3552_S1_ctrl_5_a$expected_count,
    s41_3552_S2_1p_H_octanal_5_a = s41_3552_S2_1p_H_octanal_5_a$expected_count,
    s43_3552_S4_ctrl_5_b = s43_3552_S4_ctrl_5_b$expected_count,
    s44_3552_S5_1p_H_octanal_5_b = s44_3552_S5_1p_H_octanal_5_b$expected_count,
    s46_3606_S1_ctrl_5_c = s46_3606_S1_ctrl_5_c$expected_count,
    s47_3606_S2_1p_H_octanal_5_c = s47_3606_S2_1p_H_octanal_5_c$expected_count,
    s49_3552_S7_ctrl_6_a = s49_3552_S7_ctrl_6_a$expected_count,
    s50_3552_S8_1p_pDihydrocarveol_6_a = s50_3552_S8_1p_pDihydrocarveol_6_a$expected_count,
    s51_3552_S9_1p_nDihydrocarveol_6_a = s51_3552_S9_1p_nDihydrocarveol_6_a$expected_count,
    s53_3639_S1_ctrl_6_b = s53_3639_S1_ctrl_6_b$expected_count,
    s54_3639_S2_1p_pDihydrocarveol_6_b = s54_3639_S2_1p_pDihydrocarveol_6_b$expected_count,
    s55_3639_S3_1p_nDihydrocarveol_6_b = s55_3639_S3_1p_nDihydrocarveol_6_b$expected_count,
    s57_3639_S5_ctrl_6_c = s57_3639_S5_ctrl_6_c$expected_count,
    s58_3639_S6_1p_pDihydrocarveol_6_c = s58_3639_S6_1p_pDihydrocarveol_6_c$expected_count,
    s59_3639_S7_1p_nDihydrocarveol_6_c = s59_3639_S7_1p_nDihydrocarveol_6_c$expected_count,
    s61_3606_S4_ctrl_7_a = s61_3606_S4_ctrl_7_a$expected_count,
    s62_3606_S5_1p_citronellol_7_a = s62_3606_S5_1p_citronellol_7_a$expected_count,
    s63_3606_S6_1_guaiacol_7_a = s63_3606_S6_1_guaiacol_7_a$expected_count,
    s64_3606_S7_ctrl_7_b = s64_3606_S7_ctrl_7_b$expected_count,
    s65_3606_S8_1p_citronellol_7_b = s65_3606_S8_1p_citronellol_7_b$expected_count,
    s66_3606_S9_1_guaiacol_7_b = s66_3606_S9_1_guaiacol_7_b$expected_count,
    s67_3606_S10_ctrl_7_c = s67_3606_S10_ctrl_7_c$expected_count,
    s68_3606_S11_1p_citronellol_7_c = s68_3606_S11_1p_citronellol_7_c$expected_count,
    s69_3606_S12_1_guaiacol_7_c = s69_3606_S12_1_guaiacol_7_c$expected_count,
    s70_3645_S1_ctrl_8_a = s70_3645_S1_ctrl_8_a$expected_count,
    s71_3645_S2_1p_pCarvone_8_a = s71_3645_S2_1p_pCarvone_8_a$expected_count,
    s72_3645_S3_1p_nCarvone_8_a = s72_3645_S3_1p_nCarvone_8_a$expected_count,
    s74_3645_S5_ctrl_8_b = s74_3645_S5_ctrl_8_b$expected_count,
    s75_3645_S6_1p_pCarvone_8_b = s75_3645_S6_1p_pCarvone_8_b$expected_count,
    s76_3645_S7_1p_nCarvone_8_b = s76_3645_S7_1p_nCarvone_8_b$expected_count,
    s78_3645_S9_ctrl_8_c = s78_3645_S9_ctrl_8_c$expected_count,
    s79_3645_S10_1p_pCarvone_8_c = s79_3645_S10_1p_pCarvone_8_c$expected_count,
    s80_3645_S11_1p_nCarvone_8_c = s80_3645_S11_1p_nCarvone_8_c$expected_count,
    s82_3769_S1_ctrl_9_a = s82_3769_S1_ctrl_9_a$expected_count,
    s83_3769_S2_1p_isoamylAcetate_9_a = s83_3769_S2_1p_isoamylAcetate_9_a$expected_count,
    s84_3769_S3_p1_isovalericAcid_9_a = s84_3769_S3_p1_isovalericAcid_9_a$expected_count,
    s85_3769_S4_3mM_androstenone_9_a = s85_3769_S4_3mM_androstenone_9_a$expected_count,
    s86_3769_S5_ctrl_9_b = s86_3769_S5_ctrl_9_b$expected_count,
    s87_3769_S6_1p_isoamylAcetate_9_b = s87_3769_S6_1p_isoamylAcetate_9_b$expected_count,
    s88_3769_S7_p1_isovalericAcid_9_b = s88_3769_S7_p1_isovalericAcid_9_b$expected_count,
    s89_3769_S8_3mM_androstenone_9_b = s89_3769_S8_3mM_androstenone_9_b$expected_count,
    s90_3769_S9_ctrl_9_c = s90_3769_S9_ctrl_9_c$expected_count,
    s91_3769_S10_1p_isoamylAcetate_9_c = s91_3769_S10_1p_isoamylAcetate_9_c$expected_count,
    s92_3769_S11_p1_isovalericAcid_9_c = s92_3769_S11_p1_isovalericAcid_9_c$expected_count,
    s93_3769_S12_3mM_androstenone_9_c = s93_3769_S12_3mM_androstenone_9_c$expected_count,
    s94_4258_S1_ctrl_10_a = s94_4258_S1_ctrl_10_a$expected_count,
    s96_4258_S3_1p_bCaryophyllene_10_a = s96_4258_S3_1p_bCaryophyllene_10_a$expected_count,
    s97_4258_S4_1p_ethylTiglate_10_a = s97_4258_S4_1p_ethylTiglate_10_a$expected_count,
    s98_4258_S5_ctrl_10_b = s98_4258_S5_ctrl_10_b$expected_count,
    s99_4258_S6_1p_pyridine_10_b = s99_4258_S6_1p_pyridine_10_b$expected_count,
    s100_4258_S7_1p_bCaryophyllene_10_b = s100_4258_S7_1p_bCaryophyllene_10_b$expected_count,
    s101_4258_S8_1p_methylSalicylate_10_b = s101_4258_S8_1p_methylSalicylate_10_b$expected_count,
    s102_4258_S9_ctrl_10_c = s102_4258_S9_ctrl_10_c$expected_count,
    s103_4258_S10_1p_pyridine_10_c = s103_4258_S10_1p_pyridine_10_c$expected_count,
    s104_4258_S11_1p_bCaryophyllene_10_c = s104_4258_S11_1p_bCaryophyllene_10_c$expected_count,
    s106_4259_S1_ctrl_11_a = s106_4259_S1_ctrl_11_a$expected_count,
    s107_4259_S2_1p_2heptanone_11_a = s107_4259_S2_1p_2heptanone_11_a$expected_count,
    s108_4259_S3_1p_linalool_11_a = s108_4259_S3_1p_linalool_11_a$expected_count,
    s109_4259_S4_1p_ethylButyrate_11_a = s109_4259_S4_1p_ethylButyrate_11_a$expected_count,
    s110_4259_S5_ctrl_11_b = s110_4259_S5_ctrl_11_b$expected_count,
    s111_4259_S6_1p_2heptanone_11_b = s111_4259_S6_1p_2heptanone_11_b$expected_count,
    s112_4259_S7_1p_linalool_11_b = s112_4259_S7_1p_linalool_11_b$expected_count,
    s113_4259_S8_1p_ethylButyrate_11_b = s113_4259_S8_1p_ethylButyrate_11_b$expected_count,
    s114_4259_S9_ctrl_11_c = s114_4259_S9_ctrl_11_c$expected_count,
    s116_4259_S11_1p_linalool_11_c = s116_4259_S11_1p_linalool_11_c$expected_count,
    s117_4259_S12_1p_ethylButyrate_11_c = s117_4259_S12_1p_ethylButyrate_11_c$expected_count,
    s118_4269_S1_ctrl_12_a = s118_4269_S1_ctrl_12_a$expected_count,
    s119_4269_S2_1M_pMenthol_12_a = s119_4269_S2_1M_pMenthol_12_a$expected_count,
    s120_4269_S3_1M_nMenthol_12_a = s120_4269_S3_1M_nMenthol_12_a$expected_count,
    s121_4269_S4_ctrl_12_b = s121_4269_S4_ctrl_12_b$expected_count,
    s122_4269_S5_1M_pMenthol_12_b = s122_4269_S5_1M_pMenthol_12_b$expected_count,
    s123_4269_S6_1M_nMenthol_12_b = s123_4269_S6_1M_nMenthol_12_b$expected_count,
    s124_4269_S7_ctrl_12_c = s124_4269_S7_ctrl_12_c$expected_count,
    s125_4269_S8_1M_pMenthol_12_c = s125_4269_S8_1M_pMenthol_12_c$expected_count,
    s126_4269_S9_1M_nMenthol_12_c = s126_4269_S9_1M_nMenthol_12_c$expected_count,
    s127_4269_S10_ctrl_13_c = s127_4269_S10_ctrl_13_c$expected_count,
    s128_4269_S11_1p_2heptanone_13_c = s128_4269_S11_1p_2heptanone_13_c$expected_count,
    s129_4269_S12_1p_pyridine_13_c = s129_4269_S12_1p_pyridine_13_c$expected_count,
    s130_4331_S1_ctrl_14_a = s130_4331_S1_ctrl_14_a$expected_count,
    s131_4331_S2_100p_ebFarnesene_14_a = s131_4331_S2_100p_ebFarnesene_14_a$expected_count,
    s132_4331_S3_1p_25dmp_14_a = s132_4331_S3_1p_25dmp_14_a$expected_count,
    s133_4331_S4_100uM_MTMT_14_a = s133_4331_S4_100uM_MTMT_14_a$expected_count,
    s134_4331_S5_ctrl_14_b = s134_4331_S5_ctrl_14_b$expected_count,
    s135_4331_S6_100p_ebFarnesene_14_b = s135_4331_S6_100p_ebFarnesene_14_b$expected_count,
    s136_4331_S7_1p_25dmp_14_b = s136_4331_S7_1p_25dmp_14_b$expected_count,
    s137_4331_S8_100uM_MTMT_14_b = s137_4331_S8_100uM_MTMT_14_b$expected_count,
    s138_4331_S9_ctrl_14_c = s138_4331_S9_ctrl_14_c$expected_count,
    s139_4331_S10_100p_ebFarnesene_14_c = s139_4331_S10_100p_ebFarnesene_14_c$expected_count,
    s140_4331_S11_1p_25dmp_14_c = s140_4331_S11_1p_25dmp_14_c$expected_count,
    s141_4331_S12_100uM_MTMT_14_c = s141_4331_S12_100uM_MTMT_14_c$expected_count,
    s142_4429_S5_ctrl_15_a = s142_4429_S5_ctrl_15_a$expected_count,
    s143_4429_S6_10p_p2octanol_15_a = s143_4429_S6_10p_p2octanol_15_a$expected_count,
    s144_4429_S7_10p_n2octanol_15_a = s144_4429_S7_10p_n2octanol_15_a$expected_count,
    s145_4429_S8_77p_dhb_15_a = s145_4429_S8_77p_dhb_15_a$expected_count,
    s146_4429_S9_ctrl_15_b = s146_4429_S9_ctrl_15_b$expected_count,
    s147_4429_S10_10p_p2octanol_15_b = s147_4429_S10_10p_p2octanol_15_b$expected_count,
    s148_4429_S11_10p_n2octanol_15_b = s148_4429_S11_10p_n2octanol_15_b$expected_count,
    s149_4429_S12_77p_dhb_15_b = s149_4429_S12_77p_dhb_15_b$expected_count,
    s150_4575_S1_ctrl_15_c = s150_4575_S1_ctrl_15_c$expected_count,
    s151_4998_S7_10p_p2octanol_15_c = s151_4998_S7_10p_p2octanol_15_c$expected_count,
    s152_4998_S8_10p_n2octanol_15_c = s152_4998_S8_10p_n2octanol_15_c$expected_count,
    s153_4575_S2_77p_dhb_15_c = s153_4575_S2_77p_dhb_15_c$expected_count,
    s158_4761_S1_ctrl_17_a = s158_4761_S1_ctrl_17_a$expected_count,
    s159_4761_S2_p01_sbt_17_a = s159_4761_S2_p01_sbt_17_a$expected_count,
    s160_4761_S3_1p_aPinene_17_a = s160_4761_S3_1p_aPinene_17_a$expected_count,
    s161_4761_S4_1p_bDamascone_17_a = s161_4761_S4_1p_bDamascone_17_a$expected_count,
    s162_4761_S5_ctrl_17_b = s162_4761_S5_ctrl_17_b$expected_count,
    s163_4761_S6_p01_sbt_17_b = s163_4761_S6_p01_sbt_17_b$expected_count,
    s164_4761_S7_1p_aPinene_17_b = s164_4761_S7_1p_aPinene_17_b$expected_count,
    s165_4761_S8_1p_bDamascone_17_b = s165_4761_S8_1p_bDamascone_17_b$expected_count,
    s166_4761_S9_ctrl_17_c = s166_4761_S9_ctrl_17_c$expected_count,
    s167_4761_S10_p01_sbt_17_c = s167_4761_S10_p01_sbt_17_c$expected_count,
    s168_4761_S11_1p_aPinene_17_c = s168_4761_S11_1p_aPinene_17_c$expected_count,
    s169_4761_S12_1p_bDamascone_17_c = s169_4761_S12_1p_bDamascone_17_c$expected_count,
    s170_4999_S1_ctrl_18_a = s170_4999_S1_ctrl_18_a$expected_count,
    s171_4999_S2_1p_2m2p_18_a = s171_4999_S2_1p_2m2p_18_a$expected_count,
    s173_4999_S4_1p_transCinnamaldehyde_18_a = s173_4999_S4_1p_transCinnamaldehyde_18_a$expected_count,
    s174_4999_S5_ctrl_18_b = s174_4999_S5_ctrl_18_b$expected_count,
    s175_4999_S6_1p_2m2p_18_b = s175_4999_S6_1p_2m2p_18_b$expected_count,
    s177_4999_S8_1p_transCinnamaldehyde_18_b = s177_4999_S8_1p_transCinnamaldehyde_18_b$expected_count,
    s178_4999_S9_ctrl_18_c = s178_4999_S9_ctrl_18_c$expected_count,
    s179_4999_S10_1p_2m2p_18_c = s179_4999_S10_1p_2m2p_18_c$expected_count,
    s181_4999_S12_1p_transCinnamaldehyde_18_c = s181_4999_S12_1p_transCinnamaldehyde_18_c$expected_count,
    s182_5000_S1_ctrl_19_a = s182_5000_S1_ctrl_19_a$expected_count,
    s183_5000_S2_1p_hexylTiglate_19_a = s183_5000_S2_1p_hexylTiglate_19_a$expected_count,
    s184_5000_S3_1p_2ethyl3methylPyrazine_19_a = s184_5000_S3_1p_2ethyl3methylPyrazine_19_a$expected_count,
    s185_5000_S4_1p_isopropylTiglate_19_a = s185_5000_S4_1p_isopropylTiglate_19_a$expected_count,
    s186_5000_S5_ctrl_19_b = s186_5000_S5_ctrl_19_b$expected_count,
    s187_5000_S6_1p_hexylTiglate_19_b = s187_5000_S6_1p_hexylTiglate_19_b$expected_count,
    s188_5000_S7_1p_2ethyl3methylPyrazine_19_b = s188_5000_S7_1p_2ethyl3methylPyrazine_19_b$expected_count,
    s189_5000_S8_1p_isopropylTiglate_19_b = s189_5000_S8_1p_isopropylTiglate_19_b$expected_count,
    s190_5000_S9_ctrl_19_c = s190_5000_S9_ctrl_19_c$expected_count,
    s191_5000_S10_1p_hexylTiglate_19_c = s191_5000_S10_1p_hexylTiglate_19_c$expected_count,
    s192_5000_S11_1p_2ethyl3methylPyrazine_19_c = s192_5000_S11_1p_2ethyl3methylPyrazine_19_c$expected_count,
    s193_5000_S12_1p_isopropylTiglate_19_c = s193_5000_S12_1p_isopropylTiglate_19_c$expected_count,
    s206_1_5mn1_ctrl_21_a = s206_1_5mn1_ctrl_21_a$expected_count,
    s207_1_5mn2_1p_2hac_21_a = s207_1_5mn2_1p_2hac_21_a$expected_count,
    s208_1_5mn3_1p_heptanal_21_a = s208_1_5mn3_1p_heptanal_21_a$expected_count,
    s209_1_5mn4_ctrl_21_b = s209_1_5mn4_ctrl_21_b$expected_count,
    s210_1_5mn5_1p_2hac_21_b = s210_1_5mn5_1p_2hac_21_b$expected_count,
    s211_1_5mn6_1p_heptanal_21_b = s211_1_5mn6_1p_heptanal_21_b$expected_count,
    s212_6_9mn4_ctrl_21_c = s212_6_9mn4_ctrl_21_c$expected_count,
    s213_6_9mn5_1p_2hac_21_c = s213_6_9mn5_1p_2hac_21_c$expected_count,
    s215_1_5mn7_ctrl_22_a = s215_1_5mn7_ctrl_22_a$expected_count,
    s216_1_5mn8_1p_4methylAcetophenone_22_a = s216_1_5mn8_1p_4methylAcetophenone_22_a$expected_count,
    s217_1_5mn9_1p_3methyl1butanethiol_22_a = s217_1_5mn9_1p_3methyl1butanethiol_22_a$expected_count,
    s218_1_5mn10_ctrl_22_b = s218_1_5mn10_ctrl_22_b$expected_count,
    s219_1_5mn11_1p_4methylAcetophenone_22_b = s219_1_5mn11_1p_4methylAcetophenone_22_b$expected_count,
    s220_1_5mn12_1p_3methyl1butanethiol_22_b = s220_1_5mn12_1p_3methyl1butanethiol_22_b$expected_count,
    s221_6_9mn1_ctrl_22_c = s221_6_9mn1_ctrl_22_c$expected_count,
    s222_6_9mn2_1p_4methylAcetophenone_22_c = s222_6_9mn2_1p_4methylAcetophenone_22_c$expected_count,
    s224_14_17mn1_ctrl_21a22_c = s224_14_17mn1_ctrl_21a22_c$expected_count,
    s225_14_17mn2_1p_3methyl1butanethiol_22_c = s225_14_17mn2_1p_3methyl1butanethiol_22_c$expected_count,
    s226_14_17mn3_1p_heptanal_21_c = s226_14_17mn3_1p_heptanal_21_c$expected_count,
    s227_6_9mn7_ctrl_23_a = s227_6_9mn7_ctrl_23_a$expected_count,
    s228_6_9mn8_1p_paraCresol_23_a = s228_6_9mn8_1p_paraCresol_23_a$expected_count,
    s229_6_9mn9_1p_dimethyltrisulfide_23_a = s229_6_9mn9_1p_dimethyltrisulfide_23_a$expected_count,
    s230_6_9mn10_ctrl_23_b = s230_6_9mn10_ctrl_23_b$expected_count,
    s231_6_9mn11_1p_paraCresol_23_b = s231_6_9mn11_1p_paraCresol_23_b$expected_count,
    s232_6_9mn12_1p_dimethyltrisulfide_23_b = s232_6_9mn12_1p_dimethyltrisulfide_23_b$expected_count,
    s233_10_13mn1_ctrl_23_c = s233_10_13mn1_ctrl_23_c$expected_count,
    s234_10_13mn2_1p_paraCresol_23_c = s234_10_13mn2_1p_paraCresol_23_c$expected_count,
    s235_10_13mn3_1p_dimethyltrisulfide_23_c = s235_10_13mn3_1p_dimethyltrisulfide_23_c$expected_count,
    s236_10_13mn4_ctrl_24_a = s236_10_13mn4_ctrl_24_a$expected_count,
    s237_10_13mn5_1p_butyricAcid_24_a = s237_10_13mn5_1p_butyricAcid_24_a$expected_count,
    s238_10_13mn6_1p_indole_24_a = s238_10_13mn6_1p_indole_24_a$expected_count,
    s239_10_13mn7_ctrl_24_b = s239_10_13mn7_ctrl_24_b$expected_count,
    s240_10_13mn8_1p_butyricAcid_24_b = s240_10_13mn8_1p_butyricAcid_24_b$expected_count,
    s241_10_13mn9_1p_indole_24_b = s241_10_13mn9_1p_indole_24_b$expected_count,
    s242_10_13mn10_ctrl_24_c = s242_10_13mn10_ctrl_24_c$expected_count,
    s243_10_13mn11_1p_butyricAcid_24_c = s243_10_13mn11_1p_butyricAcid_24_c$expected_count,
    s244_10_13mn12_1p_indole_24_c = s244_10_13mn12_1p_indole_24_c$expected_count,
    s245_14_17mn4_ctrl_25_a = s245_14_17mn4_ctrl_25_a$expected_count,
    s246_14_17mn5_1p_tbm_25_a = s246_14_17mn5_1p_tbm_25_a$expected_count,
    s247_14_17mn6_1p_diacetyl_25_a = s247_14_17mn6_1p_diacetyl_25_a$expected_count,
    s248_14_17mn7_ctrl_25_b = s248_14_17mn7_ctrl_25_b$expected_count,
    s249_14_17mn8_1p_tbm_25_b = s249_14_17mn8_1p_tbm_25_b$expected_count,
    s250_14_17mn9_1p_diacetyl_25_b = s250_14_17mn9_1p_diacetyl_25_b$expected_count,
    s251_14_17mn10_ctrl_25_c = s251_14_17mn10_ctrl_25_c$expected_count,
    s252_14_17mn11_1p_tbm_25_c = s252_14_17mn11_1p_tbm_25_c$expected_count,
    s253_14_17mn12_1p_diacetyl_25_c = s253_14_17mn12_1p_diacetyl_25_c$expected_count,
    s254_18_1_No_ctrl_26_a = s254_18_1_No_ctrl_26_a$expected_count,
    s255_18_2_He_1p_2hexanone_26_a = s255_18_2_He_1p_2hexanone_26_a$expected_count,
    s256_18_3_PEA_1p_2phenylethylalcohol_26_a = s256_18_3_PEA_1p_2phenylethylalcohol_26_a$expected_count,
    s257_19_1_No_ctrl_26_b = s257_19_1_No_ctrl_26_b$expected_count,
    s258_19_2_He_1p_2hexanone_26_b = s258_19_2_He_1p_2hexanone_26_b$expected_count,
    s259_19_3_PEA_1p_2phenylethylalcohol_26_b = s259_19_3_PEA_1p_2phenylethylalcohol_26_b$expected_count,
    s260_mn1_ctrl_26_c = s260_mn1_ctrl_26_c$expected_count,
    s261_mn2_1p_2hexanone_26_c = s261_mn2_1p_2hexanone_26_c$expected_count,
    s262_mn3_1p_2phenylethylalcohol_26_c = s262_mn3_1p_2phenylethylalcohol_26_c$expected_count,
    s263_mn4_ctrl_27_a = s263_mn4_ctrl_27_a$expected_count,
    s264_mn5_1p_tbm_27_a = s264_mn5_1p_tbm_27_a$expected_count,
    s265_mn6_1p_acetophenone_27_a = s265_mn6_1p_acetophenone_27_a$expected_count,
    s266_mn7_ctrl_27_b = s266_mn7_ctrl_27_b$expected_count,
    s267_mn8_1p_tbm_27_b = s267_mn8_1p_tbm_27_b$expected_count,
    s268_mn9_1p_acetophenone_27_b = s268_mn9_1p_acetophenone_27_b$expected_count,
    s269_mn10_ctrl_27_c = s269_mn10_ctrl_27_c$expected_count,
    s270_mn11_1p_tbm_27_c = s270_mn11_1p_tbm_27_c$expected_count,
    s271_mn12_1p_acetophenone_27_c = s271_mn12_1p_acetophenone_27_c$expected_count,
    s272_24_1_No_ctrl_28_a = s272_24_1_No_ctrl_28_a$expected_count,
    s273_24_2_Sk_p01_e2butene1thiol_28_a = s273_24_2_Sk_p01_e2butene1thiol_28_a$expected_count,
    s274_24_3_Cy_p01_cyclopentanethiol_28_a = s274_24_3_Cy_p01_cyclopentanethiol_28_a$expected_count,
    s275_25_1_No_ctrl_28_b = s275_25_1_No_ctrl_28_b$expected_count,
    s276_25_2_Sk_p01_e2butene1thiol_28_b = s276_25_2_Sk_p01_e2butene1thiol_28_b$expected_count,
    s277_25_3_Cy_p01_cyclopentanethiol_28_b = s277_25_3_Cy_p01_cyclopentanethiol_28_b$expected_count,
    s278_26_1_No_ctrl_28_c = s278_26_1_No_ctrl_28_c$expected_count,
    s279_26_2_Sk_p01_e2butene1thiol_28_c = s279_26_2_Sk_p01_e2butene1thiol_28_c$expected_count,
    s280_26_3_Cy_p01_cyclopentanethiol_28_c = s280_26_3_Cy_p01_cyclopentanethiol_28_c$expected_count,
    s281_27_1_No_ctrl_29_a = s281_27_1_No_ctrl_29_a$expected_count,
    s282_27_2_Ci_1p_citral_29_a = s282_27_2_Ci_1p_citral_29_a$expected_count,
    s283_27_3_An_1p_anisaldehyde_29_a = s283_27_3_An_1p_anisaldehyde_29_a$expected_count,
    s284_28_1_No_ctrl_29_b = s284_28_1_No_ctrl_29_b$expected_count,
    s285_28_2_Ci_1p_citral_29_b = s285_28_2_Ci_1p_citral_29_b$expected_count,
    s286_28_3_An_1p_anisaldehyde_29_b = s286_28_3_An_1p_anisaldehyde_29_b$expected_count,
    s287_29_1_No_ctrl_29_c = s287_29_1_No_ctrl_29_c$expected_count,
    s288_29_2_Ci_1p_citral_29_c = s288_29_2_Ci_1p_citral_29_c$expected_count,
    s289_29_3_An_1p_anisaldehyde_29_c = s289_29_3_An_1p_anisaldehyde_29_c$expected_count,
    s290_SerA4_TCCTGAGC_ctrl_30_a = s290_SerA4_TCCTGAGC_ctrl_30_a$expected_count,
    s291_SerA5_GGACTCCT_1p_acetophenone_30_a = s291_SerA5_GGACTCCT_1p_acetophenone_30_a$expected_count,
    s292_SerA6_TAGGCATG_100p_acetophenone_30_a = s292_SerA6_TAGGCATG_100p_acetophenone_30_a$expected_count,
    s293_SerB4_TCCTGAGC_ctrl_30_b = s293_SerB4_TCCTGAGC_ctrl_30_b$expected_count,
    s294_SerB5_GGACTCCT_1p_acetophenone_30_b = s294_SerB5_GGACTCCT_1p_acetophenone_30_b$expected_count,
    s295_SerB6_TAGGCATG_100p_acetophenone_30_b = s295_SerB6_TAGGCATG_100p_acetophenone_30_b$expected_count,
    s296_SerB10_CGAGGCTG_ctrl_30_c = s296_SerB10_CGAGGCTG_ctrl_30_c$expected_count,
    s297_SerB11_AAGAGGCA_1p_acetophenone_30_c = s297_SerB11_AAGAGGCA_1p_acetophenone_30_c$expected_count,
    s298_SerB12_GTAGAGGA_100p_acetophenone_30_c = s298_SerB12_GTAGAGGA_100p_acetophenone_30_c$expected_count,
    s299_SerA7_CTCTCTAC_ctrl_31_a = s299_SerA7_CTCTCTAC_ctrl_31_a$expected_count,
    s300_SerA8_CAGAGAGG_1p_tmt_31_a = s300_SerA8_CAGAGAGG_1p_tmt_31_a$expected_count,
    s301_SerA9_GCTACGCT_100p_tmt_31_a = s301_SerA9_GCTACGCT_100p_tmt_31_a$expected_count,
    s302_SerA10_CGAGGCTG_ctrl_31_b = s302_SerA10_CGAGGCTG_ctrl_31_b$expected_count,
    s303_SerA11_AAGAGGCA_1p_tmt_31_b = s303_SerA11_AAGAGGCA_1p_tmt_31_b$expected_count,
    s304_SerA12_GTAGAGGA_100p_tmt_31_b = s304_SerA12_GTAGAGGA_100p_tmt_31_b$expected_count,
    s305_SerB1_TAAGGCGA_ctrl_31_c = s305_SerB1_TAAGGCGA_ctrl_31_c$expected_count,
    s306_SerB2_CGTACTAG_1p_tmt_31_c = s306_SerB2_CGTACTAG_1p_tmt_31_c$expected_count,
    s307_SerB3_AGGCAGAA_100p_tmt_31_c = s307_SerB3_AGGCAGAA_100p_tmt_31_c$expected_count,
    s308_SerB7_CTCTCTAC_ctrl_32_a = s308_SerB7_CTCTCTAC_ctrl_32_a$expected_count,
    s309_SerB8_CAGAGAGG_1p_2m2t_32_a = s309_SerB8_CAGAGAGG_1p_2m2t_32_a$expected_count,
    s310_SerB9_GCTACGCT_100p_2m2t_32_a = s310_SerB9_GCTACGCT_100p_2m2t_32_a$expected_count,
    s311_SerC1_TAAGGCGA_ctrl_32_b = s311_SerC1_TAAGGCGA_ctrl_32_b$expected_count,
    s312_SerC2_CGTACTAG_1p_2m2t_32_b = s312_SerC2_CGTACTAG_1p_2m2t_32_b$expected_count,
    s313_SerC3_AGGCAGAA_100p_2m2t_32_b = s313_SerC3_AGGCAGAA_100p_2m2t_32_b$expected_count,
    s314_SerC4_TCCTGAGC_ctrl_32_c = s314_SerC4_TCCTGAGC_ctrl_32_c$expected_count,
    s315_SerC5_GGACTCCT_1p_2m2t_32_c = s315_SerC5_GGACTCCT_1p_2m2t_32_c$expected_count,
    s316_SerC6_TAGGCATG_100p_2m2t_32_c = s316_SerC6_TAGGCATG_100p_2m2t_32_c$expected_count,
    s317_SerC7_CTCTCTAC_ctrl_33_a = s317_SerC7_CTCTCTAC_ctrl_33_a$expected_count,
    s318_SerC8_CAGAGAGG_1p_sbt_33_a = s318_SerC8_CAGAGAGG_1p_sbt_33_a$expected_count,
    s319_SerC9_GCTACGCT_100p_sbt_33_a = s319_SerC9_GCTACGCT_100p_sbt_33_a$expected_count,
    s320_SerC10_CGAGGCTG_ctrl_33_b = s320_SerC10_CGAGGCTG_ctrl_33_b$expected_count,
    s321_SerC11_AAGAGGCA_1p_sbt_33_b = s321_SerC11_AAGAGGCA_1p_sbt_33_b$expected_count,
    s322_SerC12_GTAGAGGA_100p_sbt_33_b = s322_SerC12_GTAGAGGA_100p_sbt_33_b$expected_count,
    s323_SerD01_ctrl_33_c = s323_SerD01_ctrl_33_c$expected_count,
    s324_SerD02_1p_sbt_33_c = s324_SerD02_1p_sbt_33_c$expected_count,
    s325_SerD03_100p_sbt_33_c = s325_SerD03_100p_sbt_33_c$expected_count,
    s326_SerD04_ctrl_34_a = s326_SerD04_ctrl_34_a$expected_count,
    s327_SerD05_1p_ntmt_34_a = s327_SerD05_1p_ntmt_34_a$expected_count,
    s329_SerD07_ctrl_34_b = s329_SerD07_ctrl_34_b$expected_count,
    s330_SerD08_1p_ntmt_34_b = s330_SerD08_1p_ntmt_34_b$expected_count,
    s332_SerD10_ctrl_34_c = s332_SerD10_ctrl_34_c$expected_count,
    s333_SerD11_1p_ntmt_34_c = s333_SerD11_1p_ntmt_34_c$expected_count,
    s335_SerE1_L001_R1_001_ctrl_35_a = s335_SerE1_L001_R1_001_ctrl_35_a$expected_count,
    s336_SerE2_L001_R1_001_p01_acetophenone_35_a = s336_SerE2_L001_R1_001_p01_acetophenone_35_a$expected_count,
    s337_SerE3_L001_R1_001_p1_acetophenone_35_a = s337_SerE3_L001_R1_001_p1_acetophenone_35_a$expected_count,
    s338_SerE4_L001_R1_001_ctrl_35_b = s338_SerE4_L001_R1_001_ctrl_35_b$expected_count,
    s339_SerE5_L001_R1_001_p01_acetophenone_35_b = s339_SerE5_L001_R1_001_p01_acetophenone_35_b$expected_count,
    s340_SerE6_L001_R1_001_p1_acetophenone_35_b = s340_SerE6_L001_R1_001_p1_acetophenone_35_b$expected_count,
    s341_SerE7_L001_R1_001_ctrl_35_c = s341_SerE7_L001_R1_001_ctrl_35_c$expected_count,
    s342_SerE8_L001_R1_001_p01_acetophenone_35_c = s342_SerE8_L001_R1_001_p01_acetophenone_35_c$expected_count,
    s343_SerE9_L001_R1_001_p1_acetophenone_35_c = s343_SerE9_L001_R1_001_p1_acetophenone_35_c$expected_count,
    s344_SerE10_L001_R1_001_ctrl_10_a = s344_SerE10_L001_R1_001_ctrl_10_a$expected_count,
    s345_SerE11_L001_R1_001_1p_methylSalicylate_10_a = s345_SerE11_L001_R1_001_1p_methylSalicylate_10_a$expected_count,
    s347_SerH10_combine_ctrl_10_b = s347_SerH10_combine_ctrl_10_b$expected_count,
    s348_SerH11_combine_1p_methylSalicylate_10_b = s348_SerH11_combine_1p_methylSalicylate_10_b$expected_count,
    s350_SerG1_L001_R1_001_ctrl_36_a = s350_SerG1_L001_R1_001_ctrl_36_a$expected_count,
    s351_SerG2_L001_R1_001_p01_tmt_36_a = s351_SerG2_L001_R1_001_p01_tmt_36_a$expected_count,
    s352_SerG3_L001_R1_001_p1_tmt_36_a = s352_SerG3_L001_R1_001_p1_tmt_36_a$expected_count,
    s353_SerG4_L001_R1_001_10p_tmt_36_a = s353_SerG4_L001_R1_001_10p_tmt_36_a$expected_count,
    s354_SerG5_L001_R1_001_ctrl_36_b = s354_SerG5_L001_R1_001_ctrl_36_b$expected_count,
    s355_SerG6_L001_R1_001_p01_tmt_36_b = s355_SerG6_L001_R1_001_p01_tmt_36_b$expected_count,
    s356_SerG7_L001_R1_001_p1_tmt_36_b = s356_SerG7_L001_R1_001_p1_tmt_36_b$expected_count,
    s357_SerG8_L001_R1_001_10p_tmt_36_b = s357_SerG8_L001_R1_001_10p_tmt_36_b$expected_count,
    s358_SerG9_L001_R1_001_ctrl_36_c = s358_SerG9_L001_R1_001_ctrl_36_c$expected_count,
    s359_SerG10_L001_R1_001_p01_tmt_36_c = s359_SerG10_L001_R1_001_p01_tmt_36_c$expected_count,
    s360_SerG11_L001_R1_001_p1_tmt_36_c = s360_SerG11_L001_R1_001_p1_tmt_36_c$expected_count,
    s361_SerG12_L001_R1_001_10p_tmt_36_c = s361_SerG12_L001_R1_001_10p_tmt_36_c$expected_count,
    s362_SerH01_combine_ctrl_37_a = s362_SerH01_combine_ctrl_37_a$expected_count,
    s363_SerH02_combine_1p_pLimonene_37_a = s363_SerH02_combine_1p_pLimonene_37_a$expected_count,
    s364_SerH03_combine_1p_propionicAcid_37_a = s364_SerH03_combine_1p_propionicAcid_37_a$expected_count,
    s365_SerH04_combine_ctrl_37_b = s365_SerH04_combine_ctrl_37_b$expected_count,
    s366_SerH05_combine_1p_pLimonene_37_b = s366_SerH05_combine_1p_pLimonene_37_b$expected_count,
    s367_SerH06_combine_1p_propionicAcid_37_b = s367_SerH06_combine_1p_propionicAcid_37_b$expected_count,
    s368_SerH07_combine_ctrl_37_c = s368_SerH07_combine_ctrl_37_c$expected_count,
    s369_SerH08_combine_1p_pLimonene_37_c = s369_SerH08_combine_1p_pLimonene_37_c$expected_count,
    s370_SerH09_combine_1p_propionicAcid_37_c = s370_SerH09_combine_1p_propionicAcid_37_c$expected_count,
    s371_SerJ5_12072017_ctrl_38_a = s371_SerJ5_12072017_ctrl_38_a$expected_count,
    s372_SerJ6_12072017_1p_geraniol_38_a = s372_SerJ6_12072017_1p_geraniol_38_a$expected_count,
    s373_SerJ7_12072017_1p_bIonone_38_a = s373_SerJ7_12072017_1p_bIonone_38_a$expected_count,
    s374_SerJ8_12072017_1p_dimethylsulfide_38_a = s374_SerJ8_12072017_1p_dimethylsulfide_38_a$expected_count,
    s375_SerJ9_12072017_ctrl_38_b = s375_SerJ9_12072017_ctrl_38_b$expected_count,
    s376_SerJ10_12072017_1p_geraniol_38_b = s376_SerJ10_12072017_1p_geraniol_38_b$expected_count,
    s377_SerJ11_12072017_1p_bIonone_38_b = s377_SerJ11_12072017_1p_bIonone_38_b$expected_count,
    s378_SerJ12_12072017_1p_dimethylsulfide_38_b = s378_SerJ12_12072017_1p_dimethylsulfide_38_b$expected_count,
    s379_SerF01_combine_ctrl_38_c = s379_SerF01_combine_ctrl_38_c$expected_count,
    s380_SerF02_combine_1p_geraniol_38_c = s380_SerF02_combine_1p_geraniol_38_c$expected_count,
    s381_SerF03_combine_1p_bIonone_38_c = s381_SerF03_combine_1p_bIonone_38_c$expected_count,
    s382_SerF04_combine_1p_dimethylsulfide_38_c = s382_SerF04_combine_1p_dimethylsulfide_38_c$expected_count,
    s383_SerF05_combine_ctrl_39_a = s383_SerF05_combine_ctrl_39_a$expected_count,
    s384_SerF06_combine_1p_ethylTiglate_39_a = s384_SerF06_combine_1p_ethylTiglate_39_a$expected_count,
    s386_SerF08_combine_1p_heptanoicAcid_39_a = s386_SerF08_combine_1p_heptanoicAcid_39_a$expected_count,
    s387_SerF09_combine_ctrl_39_b = s387_SerF09_combine_ctrl_39_b$expected_count,
    s388_SerF10_combine_1p_ethylTiglate_39_b = s388_SerF10_combine_1p_ethylTiglate_39_b$expected_count,
    s390_SerF12_combine_1p_heptanoicAcid_39_b = s390_SerF12_combine_1p_heptanoicAcid_39_b$expected_count,
    s391_SerJ1_12072017_ctrl_39_c = s391_SerJ1_12072017_ctrl_39_c$expected_count,
    s392_SerJ2_12072017_1p_ethylTiglate_39_c = s392_SerJ2_12072017_1p_ethylTiglate_39_c$expected_count,
    s394_SerJ4_12072017_1p_heptanoicAcid_39_c = s394_SerJ4_12072017_1p_heptanoicAcid_39_c$expected_count,
    row.names = s1_3046_S1_ctrl_1_a$gene_id)


#write.csv(compiledDataCounts, "./../seq_preprocess/compiledExpectedCounts_allGene.csv")

#eID <- row.names(compiledDataCounts)
#
#eID_noVersion <- str_replace(eID,
#                             pattern = ".[0-9]+$",
#                             replacement = "")
#
#convert <- getBM(attributes = c('ensembl_gene_id','mgi_symbol'),
#                 filters='ensembl_gene_id',
#                 values = eID_noVersion,
#                 mart=useMart(biomart = "ensembl", dataset = "mmusculus_gene_ensembl"))
#
#mm_symbol_pre <- as.data.frame(org.Mm.egSYMBOL)
#mm_desc_pre <- as.data.frame(org.Mm.egGENENAME)
#mm_pre <- mm_symbol_pre %>%
#  full_join(mm_desc_pre, by="gene_id")
#
#write.csv(convert, "./../seq_preprocess/biomart_eID_symbol.csv")
#write.csv(mm_pre, "./../seq_preprocess/mm_symbol_name.csv")
