==> 2_GeneID_Compiler.ipynb <==
import pandas as pd
import numpy as np
import os

print(pd.__version__)
print(np.__version__)

os.chdir("./seq_preprocess")

counts_data = pd.read_csv("./compiledExpectedCounts_allGene.csv", index_col = 0)
biomart = pd.read_csv("./biomart_eID_symbol.csv", index_col = 0)
mm_name = pd.read_csv("./mm_symbol_name.csv", index_col = 0)

eID = counts_data.index.values.tolist()
eID = [i.split('.', 1)[0] for i in eID]
eID_compile = pd.DataFrame(eID, columns = ['eID'])

eID_sym_name = pd.DataFrame()

for i in range(eID_compile.shape[0]):
    eID = eID_compile.iloc[i,:].values[0]
    if len(biomart[biomart.loc[:,'ensembl_gene_id'] == eID]['mgi_symbol'].values) > 0:
        sym = biomart[biomart.loc[:,'ensembl_gene_id'] == eID]['mgi_symbol'].values[0]
    else:
        sym = 'no_symbol'
    if len(mm_name[mm_name['symbol'] == sym]['gene_name'].values) > 0:
        name = mm_name[mm_name['symbol'] == sym]['gene_name'].values[0]
    else:
        name = 'no_name'
    eID_sym_name = pd.concat([eID_sym_name, pd.DataFrame([eID, sym, name]).transpose()], axis = 0)
    

#eID_sym_name.to_csv("./eID_sym_name.csv")
eID_sym_name = pd.read_csv("./eID_sym_name.csv", index_col = 0)
