==> 3_DifferentialExpression.ipynb <==
library(tidyverse)
library(edgeR)
library(stringr)

print(sessionInfo())

setwd("./seq_preprocess/")

geneID_info <- read.delim("./eID_sym_name.csv", header=T, sep=',')
geneID_info <- geneID_info[,c(2,3,4)]
colnames(geneID_info) <- c("ensembl_gene_id","symbol","name")

all_expectedCountsTable <- read.delim("./compiledExpectedCounts_allGene.csv", row.names = 1, header=T, sep=',')
meta_ST3 <- data.frame(FileName=c("No1","No2","No3","ST1","ST2","ST3"),Subject=c("A","B","C","A","B","C"),Treatment=c(0,0,0,1,1,1))

pS6_countsTable_100mM_CPD <- round(data.frame(
    No1=all_expectedCountsTable$s1_3046_S1_ctrl_1_a,
    No2=all_expectedCountsTable$s4_3046_S4_ctrl_1_b,
    No3=all_expectedCountsTable$s7_3134_S4_ctrl_1_c,
    ST1=all_expectedCountsTable$s2_3046_S2_100mM_H_CPD_1_a,
    ST2=all_expectedCountsTable$s5_3046_S5_100mM_H_CPD_1_b,
    ST3=all_expectedCountsTable$s8_3134_S5_100mM_H_CPD_1_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_10mM_MTMT <- round(data.frame(
    No1=all_expectedCountsTable$s10_3046_S7_ctrl_2_a,
    No2=all_expectedCountsTable$s13_3046_S10_ctrl_2_b,
    No3=all_expectedCountsTable$s16_3134_S1_ctrl_2_c,
    ST1=all_expectedCountsTable$s11_3046_S8_10mM_H_MTMT_2_a,
    ST2=all_expectedCountsTable$s14_3046_S11_10mM_H_MTMT_2_b,
    ST3=all_expectedCountsTable$s17_3134_S2_10mM_H_MTMT_2_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_Benzaldehyde <- round(data.frame(
    No1=all_expectedCountsTable$s19_3134_S7_ctrl_3_a,
    No2=all_expectedCountsTable$s22_3134_S10_ctrl_3_b,
    No3=all_expectedCountsTable$s25_3284_S10_ctrl_3_c,
    ST1=all_expectedCountsTable$s20_3134_S8_1p_H_benzaldehyde_3_a,
    ST2=all_expectedCountsTable$s23_3134_S11_1p_H_benzaldehyde_3_b,
    ST3=all_expectedCountsTable$s26_3284_S11_1p_H_benzaldehyde_3_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_Octanal <- round(data.frame(
    No1=all_expectedCountsTable$s40_3552_S1_ctrl_5_a,
    No2=all_expectedCountsTable$s43_3552_S4_ctrl_5_b,
    No3=all_expectedCountsTable$s46_3606_S1_ctrl_5_c,
    ST1=all_expectedCountsTable$s41_3552_S2_1p_H_octanal_5_a,
    ST2=all_expectedCountsTable$s44_3552_S5_1p_H_octanal_5_b,
    ST3=all_expectedCountsTable$s47_3606_S2_1p_H_octanal_5_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_pDihydrocarveol <- round(data.frame(
    No1=all_expectedCountsTable$s49_3552_S7_ctrl_6_a,
    No2=all_expectedCountsTable$s53_3639_S1_ctrl_6_b,
    No3=all_expectedCountsTable$s57_3639_S5_ctrl_6_c,
    ST1=all_expectedCountsTable$s50_3552_S8_1p_pDihydrocarveol_6_a,
    ST2=all_expectedCountsTable$s54_3639_S2_1p_pDihydrocarveol_6_b,
    ST3=all_expectedCountsTable$s58_3639_S6_1p_pDihydrocarveol_6_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_nDihydrocarveol <- round(data.frame(
    No1=all_expectedCountsTable$s49_3552_S7_ctrl_6_a,
    No2=all_expectedCountsTable$s53_3639_S1_ctrl_6_b,
    No3=all_expectedCountsTable$s57_3639_S5_ctrl_6_c,
    ST1=all_expectedCountsTable$s51_3552_S9_1p_nDihydrocarveol_6_a,
    ST2=all_expectedCountsTable$s55_3639_S3_1p_nDihydrocarveol_6_b,
    ST3=all_expectedCountsTable$s59_3639_S7_1p_nDihydrocarveol_6_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_citronellol <- round(data.frame(
    No1=all_expectedCountsTable$s61_3606_S4_ctrl_7_a, 
    No2=all_expectedCountsTable$s64_3606_S7_ctrl_7_b,
    No3=all_expectedCountsTable$s67_3606_S10_ctrl_7_c,
    ST1=all_expectedCountsTable$s62_3606_S5_1p_citronellol_7_a,
    ST2=all_expectedCountsTable$s65_3606_S8_1p_citronellol_7_b,
    ST3=all_expectedCountsTable$s68_3606_S11_1p_citronellol_7_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_guaiacol <- round(data.frame(
    No1=all_expectedCountsTable$s61_3606_S4_ctrl_7_a, 
    No2=all_expectedCountsTable$s64_3606_S7_ctrl_7_b,
    No3=all_expectedCountsTable$s67_3606_S10_ctrl_7_c,
    ST1=all_expectedCountsTable$s63_3606_S6_1_guaiacol_7_a,
    ST2=all_expectedCountsTable$s66_3606_S9_1_guaiacol_7_b,
    ST3=all_expectedCountsTable$s69_3606_S12_1_guaiacol_7_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_pCarvone <- round(data.frame(
    No1=all_expectedCountsTable$s70_3645_S1_ctrl_8_a, 
    No2=all_expectedCountsTable$s74_3645_S5_ctrl_8_b,
    No3=all_expectedCountsTable$s78_3645_S9_ctrl_8_c,
    ST1=all_expectedCountsTable$s71_3645_S2_1p_pCarvone_8_a,
    ST2=all_expectedCountsTable$s75_3645_S6_1p_pCarvone_8_b,
    ST3=all_expectedCountsTable$s79_3645_S10_1p_pCarvone_8_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_nCarvone <- round(data.frame(
    No1=all_expectedCountsTable$s70_3645_S1_ctrl_8_a, 
    No2=all_expectedCountsTable$s74_3645_S5_ctrl_8_b,
    No3=all_expectedCountsTable$s78_3645_S9_ctrl_8_c,
    ST1=all_expectedCountsTable$s72_3645_S3_1p_nCarvone_8_a,
    ST2=all_expectedCountsTable$s76_3645_S7_1p_nCarvone_8_b,
    ST3=all_expectedCountsTable$s80_3645_S11_1p_nCarvone_8_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_isoamylAcetate <- round(data.frame(
    No1=all_expectedCountsTable$s82_3769_S1_ctrl_9_a, 
    No2=all_expectedCountsTable$s86_3769_S5_ctrl_9_b,
    No3=all_expectedCountsTable$s90_3769_S9_ctrl_9_c,
    ST1=all_expectedCountsTable$s83_3769_S2_1p_isoamylAcetate_9_a,
    ST2=all_expectedCountsTable$s87_3769_S6_1p_isoamylAcetate_9_b,
    ST3=all_expectedCountsTable$s91_3769_S10_1p_isoamylAcetate_9_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_p1_isovalericAcid <- round(data.frame(
    No1=all_expectedCountsTable$s82_3769_S1_ctrl_9_a, 
    No2=all_expectedCountsTable$s86_3769_S5_ctrl_9_b,
    No3=all_expectedCountsTable$s90_3769_S9_ctrl_9_c,
    ST1=all_expectedCountsTable$s84_3769_S3_p1_isovalericAcid_9_a,
    ST2=all_expectedCountsTable$s88_3769_S7_p1_isovalericAcid_9_b,
    ST3=all_expectedCountsTable$s92_3769_S11_p1_isovalericAcid_9_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_3mM_androstenone <- round(data.frame(
    No1=all_expectedCountsTable$s82_3769_S1_ctrl_9_a, 
    No2=all_expectedCountsTable$s86_3769_S5_ctrl_9_b,
    No3=all_expectedCountsTable$s90_3769_S9_ctrl_9_c,
    ST1=all_expectedCountsTable$s85_3769_S4_3mM_androstenone_9_a,
    ST2=all_expectedCountsTable$s89_3769_S8_3mM_androstenone_9_b,
    ST3=all_expectedCountsTable$s93_3769_S12_3mM_androstenone_9_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_bCaryophyllene <- round(data.frame(
    No1=all_expectedCountsTable$s94_4258_S1_ctrl_10_a, 
    No2=all_expectedCountsTable$s98_4258_S5_ctrl_10_b,
    No3=all_expectedCountsTable$s102_4258_S9_ctrl_10_c,
    ST1=all_expectedCountsTable$s96_4258_S3_1p_bCaryophyllene_10_a,
    ST2=all_expectedCountsTable$s100_4258_S7_1p_bCaryophyllene_10_b,
    ST3=all_expectedCountsTable$s104_4258_S11_1p_bCaryophyllene_10_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_pyridine <- round(data.frame(
    No1=all_expectedCountsTable$s102_4258_S9_ctrl_10_c, 
    No2=all_expectedCountsTable$s98_4258_S5_ctrl_10_b,
    No3=all_expectedCountsTable$s127_4269_S10_ctrl_13_c,
    ST1=all_expectedCountsTable$s103_4258_S10_1p_pyridine_10_c,
    ST2=all_expectedCountsTable$s99_4258_S6_1p_pyridine_10_b,
    ST3=all_expectedCountsTable$s129_4269_S12_1p_pyridine_13_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_methylSalicylate <- round(data.frame(
    No1=all_expectedCountsTable$s98_4258_S5_ctrl_10_b, 
    No2=all_expectedCountsTable$s344_SerE10_L001_R1_001_ctrl_10_a,
    No3=all_expectedCountsTable$s347_SerH10_combine_ctrl_10_b,
    ST1=all_expectedCountsTable$s101_4258_S8_1p_methylSalicylate_10_b,
    ST2=all_expectedCountsTable$s345_SerE11_L001_R1_001_1p_methylSalicylate_10_a,
    ST3=all_expectedCountsTable$s348_SerH11_combine_1p_methylSalicylate_10_b,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_2heptanone <- round(data.frame(
    No1=all_expectedCountsTable$s106_4259_S1_ctrl_11_a, 
    No2=all_expectedCountsTable$s110_4259_S5_ctrl_11_b,
    No3=all_expectedCountsTable$s127_4269_S10_ctrl_13_c,
    ST1=all_expectedCountsTable$s107_4259_S2_1p_2heptanone_11_a,
    ST2=all_expectedCountsTable$s111_4259_S6_1p_2heptanone_11_b,
    ST3=all_expectedCountsTable$s128_4269_S11_1p_2heptanone_13_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_linalool <- round(data.frame(
    No1=all_expectedCountsTable$s106_4259_S1_ctrl_11_a, 
    No2=all_expectedCountsTable$s110_4259_S5_ctrl_11_b,
    No3=all_expectedCountsTable$s114_4259_S9_ctrl_11_c,
    ST1=all_expectedCountsTable$s108_4259_S3_1p_linalool_11_a,
    ST2=all_expectedCountsTable$s112_4259_S7_1p_linalool_11_b,
    ST3=all_expectedCountsTable$s116_4259_S11_1p_linalool_11_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_ethylButyrate <- round(data.frame(
    No1=all_expectedCountsTable$s106_4259_S1_ctrl_11_a, 
    No2=all_expectedCountsTable$s110_4259_S5_ctrl_11_b,
    No3=all_expectedCountsTable$s114_4259_S9_ctrl_11_c,
    ST1=all_expectedCountsTable$s109_4259_S4_1p_ethylButyrate_11_a,
    ST2=all_expectedCountsTable$s113_4259_S8_1p_ethylButyrate_11_b,
    ST3=all_expectedCountsTable$s117_4259_S12_1p_ethylButyrate_11_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1M_pMenthol <- round(data.frame(
    No1=all_expectedCountsTable$s118_4269_S1_ctrl_12_a, 
    No2=all_expectedCountsTable$s121_4269_S4_ctrl_12_b,
    No3=all_expectedCountsTable$s124_4269_S7_ctrl_12_c,
    ST1=all_expectedCountsTable$s119_4269_S2_1M_pMenthol_12_a,
    ST2=all_expectedCountsTable$s122_4269_S5_1M_pMenthol_12_b,
    ST3=all_expectedCountsTable$s125_4269_S8_1M_pMenthol_12_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1M_nMenthol <- round(data.frame(
    No1=all_expectedCountsTable$s118_4269_S1_ctrl_12_a, 
    No2=all_expectedCountsTable$s121_4269_S4_ctrl_12_b,
    No3=all_expectedCountsTable$s124_4269_S7_ctrl_12_c,
    ST1=all_expectedCountsTable$s120_4269_S3_1M_nMenthol_12_a,
    ST2=all_expectedCountsTable$s123_4269_S6_1M_nMenthol_12_b,
    ST3=all_expectedCountsTable$s126_4269_S9_1M_nMenthol_12_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_100p_ebFarnesene <- round(data.frame(
    No1=all_expectedCountsTable$s130_4331_S1_ctrl_14_a, 
    No2=all_expectedCountsTable$s134_4331_S5_ctrl_14_b,
    No3=all_expectedCountsTable$s138_4331_S9_ctrl_14_c,
    ST1=all_expectedCountsTable$s131_4331_S2_100p_ebFarnesene_14_a,
    ST2=all_expectedCountsTable$s135_4331_S6_100p_ebFarnesene_14_b,
    ST3=all_expectedCountsTable$s139_4331_S10_100p_ebFarnesene_14_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_25dmp <- round(data.frame(
    No1=all_expectedCountsTable$s130_4331_S1_ctrl_14_a, 
    No2=all_expectedCountsTable$s134_4331_S5_ctrl_14_b,
    No3=all_expectedCountsTable$s138_4331_S9_ctrl_14_c,
    ST1=all_expectedCountsTable$s132_4331_S3_1p_25dmp_14_a,
    ST2=all_expectedCountsTable$s136_4331_S7_1p_25dmp_14_b,
    ST3=all_expectedCountsTable$s140_4331_S11_1p_25dmp_14_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_100uM_MTMT <- round(data.frame(
    No1=all_expectedCountsTable$s130_4331_S1_ctrl_14_a, 
    No2=all_expectedCountsTable$s134_4331_S5_ctrl_14_b,
    No3=all_expectedCountsTable$s138_4331_S9_ctrl_14_c,
    ST1=all_expectedCountsTable$s133_4331_S4_100uM_MTMT_14_a,
    ST2=all_expectedCountsTable$s137_4331_S8_100uM_MTMT_14_b,
    ST3=all_expectedCountsTable$s141_4331_S12_100uM_MTMT_14_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_10p_p2octanol <- round(data.frame(
    No1=all_expectedCountsTable$s142_4429_S5_ctrl_15_a, 
    No2=all_expectedCountsTable$s146_4429_S9_ctrl_15_b,
    No3=all_expectedCountsTable$s150_4575_S1_ctrl_15_c,
    ST1=all_expectedCountsTable$s143_4429_S6_10p_p2octanol_15_a,
    ST2=all_expectedCountsTable$s147_4429_S10_10p_p2octanol_15_b,
    ST3=all_expectedCountsTable$s151_4998_S7_10p_p2octanol_15_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_10p_n2octanol <- round(data.frame(
    No1=all_expectedCountsTable$s142_4429_S5_ctrl_15_a, 
    No2=all_expectedCountsTable$s146_4429_S9_ctrl_15_b,
    No3=all_expectedCountsTable$s150_4575_S1_ctrl_15_c,
    ST1=all_expectedCountsTable$s144_4429_S7_10p_n2octanol_15_a,
    ST2=all_expectedCountsTable$s148_4429_S11_10p_n2octanol_15_b,
    ST3=all_expectedCountsTable$s152_4998_S8_10p_n2octanol_15_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_77p_dhb <- round(data.frame(
    No1=all_expectedCountsTable$s142_4429_S5_ctrl_15_a, 
    No2=all_expectedCountsTable$s146_4429_S9_ctrl_15_b,
    No3=all_expectedCountsTable$s150_4575_S1_ctrl_15_c,
    ST1=all_expectedCountsTable$s145_4429_S8_77p_dhb_15_a,
    ST2=all_expectedCountsTable$s149_4429_S12_77p_dhb_15_b,
    ST3=all_expectedCountsTable$s153_4575_S2_77p_dhb_15_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_p01_sbt <- round(data.frame(
    No1=all_expectedCountsTable$s158_4761_S1_ctrl_17_a, 
    No2=all_expectedCountsTable$s162_4761_S5_ctrl_17_b,
    No3=all_expectedCountsTable$s166_4761_S9_ctrl_17_c,
    ST1=all_expectedCountsTable$s159_4761_S2_p01_sbt_17_a,
    ST2=all_expectedCountsTable$s163_4761_S6_p01_sbt_17_b,
    ST3=all_expectedCountsTable$s167_4761_S10_p01_sbt_17_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_aPinene <- round(data.frame(
    No1=all_expectedCountsTable$s158_4761_S1_ctrl_17_a, 
    No2=all_expectedCountsTable$s162_4761_S5_ctrl_17_b,
    No3=all_expectedCountsTable$s166_4761_S9_ctrl_17_c,
    ST1=all_expectedCountsTable$s160_4761_S3_1p_aPinene_17_a,
    ST2=all_expectedCountsTable$s164_4761_S7_1p_aPinene_17_b,
    ST3=all_expectedCountsTable$s168_4761_S11_1p_aPinene_17_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_bDamascone <- round(data.frame(
    No1=all_expectedCountsTable$s158_4761_S1_ctrl_17_a, 
    No2=all_expectedCountsTable$s162_4761_S5_ctrl_17_b,
    No3=all_expectedCountsTable$s166_4761_S9_ctrl_17_c,
    ST1=all_expectedCountsTable$s161_4761_S4_1p_bDamascone_17_a,
    ST2=all_expectedCountsTable$s165_4761_S8_1p_bDamascone_17_b,
    ST3=all_expectedCountsTable$s169_4761_S12_1p_bDamascone_17_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_2m2p <- round(data.frame(
    No1=all_expectedCountsTable$s170_4999_S1_ctrl_18_a, 
    No2=all_expectedCountsTable$s174_4999_S5_ctrl_18_b,
    No3=all_expectedCountsTable$s178_4999_S9_ctrl_18_c,
    ST1=all_expectedCountsTable$s171_4999_S2_1p_2m2p_18_a,
    ST2=all_expectedCountsTable$s175_4999_S6_1p_2m2p_18_b,
    ST3=all_expectedCountsTable$s179_4999_S10_1p_2m2p_18_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_transCinnamaldehyde <- round(data.frame(
    No1=all_expectedCountsTable$s170_4999_S1_ctrl_18_a, 
    No2=all_expectedCountsTable$s174_4999_S5_ctrl_18_b,
    No3=all_expectedCountsTable$s178_4999_S9_ctrl_18_c,
    ST1=all_expectedCountsTable$s173_4999_S4_1p_transCinnamaldehyde_18_a,
    ST2=all_expectedCountsTable$s177_4999_S8_1p_transCinnamaldehyde_18_b,
    ST3=all_expectedCountsTable$s181_4999_S12_1p_transCinnamaldehyde_18_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_hexylTiglate <- round(data.frame(
    No1=all_expectedCountsTable$s182_5000_S1_ctrl_19_a, 
    No2=all_expectedCountsTable$s186_5000_S5_ctrl_19_b,
    No3=all_expectedCountsTable$s190_5000_S9_ctrl_19_c,
    ST1=all_expectedCountsTable$s183_5000_S2_1p_hexylTiglate_19_a,
    ST2=all_expectedCountsTable$s187_5000_S6_1p_hexylTiglate_19_b,
    ST3=all_expectedCountsTable$s191_5000_S10_1p_hexylTiglate_19_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_2e3mp <- round(data.frame(
    No1=all_expectedCountsTable$s182_5000_S1_ctrl_19_a, 
    No2=all_expectedCountsTable$s186_5000_S5_ctrl_19_b,
    No3=all_expectedCountsTable$s190_5000_S9_ctrl_19_c,
    ST1=all_expectedCountsTable$s184_5000_S3_1p_2ethyl3methylPyrazine_19_a,
    ST2=all_expectedCountsTable$s188_5000_S7_1p_2ethyl3methylPyrazine_19_b,
    ST3=all_expectedCountsTable$s192_5000_S11_1p_2ethyl3methylPyrazine_19_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_isopropylTiglate <- round(data.frame(
    No1=all_expectedCountsTable$s182_5000_S1_ctrl_19_a, 
    No2=all_expectedCountsTable$s186_5000_S5_ctrl_19_b,
    No3=all_expectedCountsTable$s190_5000_S9_ctrl_19_c,
    ST1=all_expectedCountsTable$s185_5000_S4_1p_isopropylTiglate_19_a,
    ST2=all_expectedCountsTable$s189_5000_S8_1p_isopropylTiglate_19_b,
    ST3=all_expectedCountsTable$s193_5000_S12_1p_isopropylTiglate_19_c,
    row.names=geneID_info$ensembl_gene_id))


pS6_countsTable_1p_2hac <- round(data.frame(
    No1=all_expectedCountsTable$s206_1_5mn1_ctrl_21_a, 
    No2=all_expectedCountsTable$s209_1_5mn4_ctrl_21_b,
    No3=all_expectedCountsTable$s212_6_9mn4_ctrl_21_c,
    ST1=all_expectedCountsTable$s207_1_5mn2_1p_2hac_21_a,
    ST2=all_expectedCountsTable$s210_1_5mn5_1p_2hac_21_b,
    ST3=all_expectedCountsTable$s213_6_9mn5_1p_2hac_21_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_heptanal <- round(data.frame(
    No1=all_expectedCountsTable$s206_1_5mn1_ctrl_21_a, 
    No2=all_expectedCountsTable$s209_1_5mn4_ctrl_21_b,
    No3=all_expectedCountsTable$s224_14_17mn1_ctrl_21a22_c,
    ST1=all_expectedCountsTable$s208_1_5mn3_1p_heptanal_21_a,
    ST2=all_expectedCountsTable$s211_1_5mn6_1p_heptanal_21_b,
    ST3=all_expectedCountsTable$s226_14_17mn3_1p_heptanal_21_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_4methylAC <- round(data.frame(
    No1=all_expectedCountsTable$s215_1_5mn7_ctrl_22_a, 
    No2=all_expectedCountsTable$s218_1_5mn10_ctrl_22_b,
    No3=all_expectedCountsTable$s221_6_9mn1_ctrl_22_c,
    ST1=all_expectedCountsTable$s216_1_5mn8_1p_4methylAcetophenone_22_a,
    ST2=all_expectedCountsTable$s219_1_5mn11_1p_4methylAcetophenone_22_b,
    ST3=all_expectedCountsTable$s222_6_9mn2_1p_4methylAcetophenone_22_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_3methyl1butanethiol <- round(data.frame(
    No1=all_expectedCountsTable$s215_1_5mn7_ctrl_22_a, 
    No2=all_expectedCountsTable$s218_1_5mn10_ctrl_22_b,
    No3=all_expectedCountsTable$s224_14_17mn1_ctrl_21a22_c,
    ST1=all_expectedCountsTable$s217_1_5mn9_1p_3methyl1butanethiol_22_a,
    ST2=all_expectedCountsTable$s220_1_5mn12_1p_3methyl1butanethiol_22_b,
    ST3=all_expectedCountsTable$s225_14_17mn2_1p_3methyl1butanethiol_22_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_paraCresol <- round(data.frame(
    No1=all_expectedCountsTable$s227_6_9mn7_ctrl_23_a, 
    No2=all_expectedCountsTable$s230_6_9mn10_ctrl_23_b,
    No3=all_expectedCountsTable$s233_10_13mn1_ctrl_23_c,
    ST1=all_expectedCountsTable$s228_6_9mn8_1p_paraCresol_23_a,
    ST2=all_expectedCountsTable$s231_6_9mn11_1p_paraCresol_23_b,
    ST3=all_expectedCountsTable$s234_10_13mn2_1p_paraCresol_23_c,
    row.names=geneID_info$ensembl_gene_id))


pS6_countsTable_1p_dimethyltrisulfide <- round(data.frame(
    No1=all_expectedCountsTable$s227_6_9mn7_ctrl_23_a, 
    No2=all_expectedCountsTable$s230_6_9mn10_ctrl_23_b,
    No3=all_expectedCountsTable$s233_10_13mn1_ctrl_23_c,
    ST1=all_expectedCountsTable$s229_6_9mn9_1p_dimethyltrisulfide_23_a,
    ST2=all_expectedCountsTable$s232_6_9mn12_1p_dimethyltrisulfide_23_b,
    ST3=all_expectedCountsTable$s235_10_13mn3_1p_dimethyltrisulfide_23_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_butyricAcid <- round(data.frame(
    No1=all_expectedCountsTable$s236_10_13mn4_ctrl_24_a, 
    No2=all_expectedCountsTable$s239_10_13mn7_ctrl_24_b,
    No3=all_expectedCountsTable$s242_10_13mn10_ctrl_24_c,
    ST1=all_expectedCountsTable$s237_10_13mn5_1p_butyricAcid_24_a,
    ST2=all_expectedCountsTable$s240_10_13mn8_1p_butyricAcid_24_b,
    ST3=all_expectedCountsTable$s243_10_13mn11_1p_butyricAcid_24_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_indole <- round(data.frame(
    No1=all_expectedCountsTable$s236_10_13mn4_ctrl_24_a, 
    No2=all_expectedCountsTable$s239_10_13mn7_ctrl_24_b,
    No3=all_expectedCountsTable$s242_10_13mn10_ctrl_24_c,
    ST1=all_expectedCountsTable$s238_10_13mn6_1p_indole_24_a,
    ST2=all_expectedCountsTable$s241_10_13mn9_1p_indole_24_b,
    ST3=all_expectedCountsTable$s244_10_13mn12_1p_indole_24_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_tbm <- round(data.frame(
    No1=all_expectedCountsTable$s245_14_17mn4_ctrl_25_a, 
    No2=all_expectedCountsTable$s248_14_17mn7_ctrl_25_b,
    No3=all_expectedCountsTable$s251_14_17mn10_ctrl_25_c,
    ST1=all_expectedCountsTable$s246_14_17mn5_1p_tbm_25_a,
    ST2=all_expectedCountsTable$s249_14_17mn8_1p_tbm_25_b,
    ST3=all_expectedCountsTable$s252_14_17mn11_1p_tbm_25_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_diacetyl <- round(data.frame(
    No1=all_expectedCountsTable$s245_14_17mn4_ctrl_25_a, 
    No2=all_expectedCountsTable$s248_14_17mn7_ctrl_25_b,
    No3=all_expectedCountsTable$s251_14_17mn10_ctrl_25_c,
    ST1=all_expectedCountsTable$s247_14_17mn6_1p_diacetyl_25_a,
    ST2=all_expectedCountsTable$s250_14_17mn9_1p_diacetyl_25_b,
    ST3=all_expectedCountsTable$s253_14_17mn12_1p_diacetyl_25_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_2hexanone <- round(data.frame(
    No1=all_expectedCountsTable$s254_18_1_No_ctrl_26_a, 
    No2=all_expectedCountsTable$s257_19_1_No_ctrl_26_b,
    No3=all_expectedCountsTable$s260_mn1_ctrl_26_c,
    ST1=all_expectedCountsTable$s255_18_2_He_1p_2hexanone_26_a,
    ST2=all_expectedCountsTable$s258_19_2_He_1p_2hexanone_26_b,
    ST3=all_expectedCountsTable$s261_mn2_1p_2hexanone_26_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_2phenylAlcohol <- round(data.frame(
    No1=all_expectedCountsTable$s254_18_1_No_ctrl_26_a, 
    No2=all_expectedCountsTable$s257_19_1_No_ctrl_26_b,
    No3=all_expectedCountsTable$s260_mn1_ctrl_26_c,
    ST1=all_expectedCountsTable$s256_18_3_PEA_1p_2phenylethylalcohol_26_a,
    ST2=all_expectedCountsTable$s259_19_3_PEA_1p_2phenylethylalcohol_26_b,
    ST3=all_expectedCountsTable$s262_mn3_1p_2phenylethylalcohol_26_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_p01_e2butene1thiol <- round(data.frame(
    No1=all_expectedCountsTable$s272_24_1_No_ctrl_28_a, 
    No2=all_expectedCountsTable$s275_25_1_No_ctrl_28_b,
    No3=all_expectedCountsTable$s278_26_1_No_ctrl_28_c,
    ST1=all_expectedCountsTable$s273_24_2_Sk_p01_e2butene1thiol_28_a,
    ST2=all_expectedCountsTable$s276_25_2_Sk_p01_e2butene1thiol_28_b,
    ST3=all_expectedCountsTable$s279_26_2_Sk_p01_e2butene1thiol_28_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_p01_cyclopentanethiol <- round(data.frame(
    No1=all_expectedCountsTable$s272_24_1_No_ctrl_28_a, 
    No2=all_expectedCountsTable$s275_25_1_No_ctrl_28_b,
    No3=all_expectedCountsTable$s278_26_1_No_ctrl_28_c,
    ST1=all_expectedCountsTable$s274_24_3_Cy_p01_cyclopentanethiol_28_a,
    ST2=all_expectedCountsTable$s277_25_3_Cy_p01_cyclopentanethiol_28_b,
    ST3=all_expectedCountsTable$s280_26_3_Cy_p01_cyclopentanethiol_28_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_citral <- round(data.frame(
    No1=all_expectedCountsTable$s281_27_1_No_ctrl_29_a, 
    No2=all_expectedCountsTable$s284_28_1_No_ctrl_29_b,
    No3=all_expectedCountsTable$s287_29_1_No_ctrl_29_c,
    ST1=all_expectedCountsTable$s282_27_2_Ci_1p_citral_29_a,
    ST2=all_expectedCountsTable$s285_28_2_Ci_1p_citral_29_b,
    ST3=all_expectedCountsTable$s288_29_2_Ci_1p_citral_29_c,
    row.names=geneID_info$ensembl_gene_id))


pS6_countsTable_1p_anisaldehyde <- round(data.frame(
    No1=all_expectedCountsTable$s281_27_1_No_ctrl_29_a, 
    No2=all_expectedCountsTable$s284_28_1_No_ctrl_29_b,
    No3=all_expectedCountsTable$s287_29_1_No_ctrl_29_c,
    ST1=all_expectedCountsTable$s283_27_3_An_1p_anisaldehyde_29_a,
    ST2=all_expectedCountsTable$s286_28_3_An_1p_anisaldehyde_29_b,
    ST3=all_expectedCountsTable$s289_29_3_An_1p_anisaldehyde_29_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_acetophenone <- round(data.frame(
    No1=all_expectedCountsTable$s290_SerA4_TCCTGAGC_ctrl_30_a, 
    No2=all_expectedCountsTable$s293_SerB4_TCCTGAGC_ctrl_30_b,
    No3=all_expectedCountsTable$s296_SerB10_CGAGGCTG_ctrl_30_c,
    ST1=all_expectedCountsTable$s291_SerA5_GGACTCCT_1p_acetophenone_30_a,
    ST2=all_expectedCountsTable$s294_SerB5_GGACTCCT_1p_acetophenone_30_b,
    ST3=all_expectedCountsTable$s297_SerB11_AAGAGGCA_1p_acetophenone_30_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_100p_acetophenone <- round(data.frame(
    No1=all_expectedCountsTable$s290_SerA4_TCCTGAGC_ctrl_30_a, 
    No2=all_expectedCountsTable$s293_SerB4_TCCTGAGC_ctrl_30_b,
    No3=all_expectedCountsTable$s296_SerB10_CGAGGCTG_ctrl_30_c,
    ST1=all_expectedCountsTable$s292_SerA6_TAGGCATG_100p_acetophenone_30_a,
    ST2=all_expectedCountsTable$s295_SerB6_TAGGCATG_100p_acetophenone_30_b,
    ST3=all_expectedCountsTable$s298_SerB12_GTAGAGGA_100p_acetophenone_30_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_tmt <- round(data.frame(
    No1=all_expectedCountsTable$s299_SerA7_CTCTCTAC_ctrl_31_a, 
    No2=all_expectedCountsTable$s302_SerA10_CGAGGCTG_ctrl_31_b,
    No3=all_expectedCountsTable$s305_SerB1_TAAGGCGA_ctrl_31_c,
    ST1=all_expectedCountsTable$s300_SerA8_CAGAGAGG_1p_tmt_31_a,
    ST2=all_expectedCountsTable$s303_SerA11_AAGAGGCA_1p_tmt_31_b,
    ST3=all_expectedCountsTable$s306_SerB2_CGTACTAG_1p_tmt_31_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_100p_tmt <- round(data.frame(
    No1=all_expectedCountsTable$s299_SerA7_CTCTCTAC_ctrl_31_a, 
    No2=all_expectedCountsTable$s302_SerA10_CGAGGCTG_ctrl_31_b,
    No3=all_expectedCountsTable$s305_SerB1_TAAGGCGA_ctrl_31_c,
    ST1=all_expectedCountsTable$s301_SerA9_GCTACGCT_100p_tmt_31_a,
    ST2=all_expectedCountsTable$s304_SerA12_GTAGAGGA_100p_tmt_31_b,
    ST3=all_expectedCountsTable$s307_SerB3_AGGCAGAA_100p_tmt_31_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_2m2t <- round(data.frame(
    No1=all_expectedCountsTable$s308_SerB7_CTCTCTAC_ctrl_32_a, 
    No2=all_expectedCountsTable$s311_SerC1_TAAGGCGA_ctrl_32_b,
    No3=all_expectedCountsTable$s314_SerC4_TCCTGAGC_ctrl_32_c,
    ST1=all_expectedCountsTable$s309_SerB8_CAGAGAGG_1p_2m2t_32_a,
    ST2=all_expectedCountsTable$s312_SerC2_CGTACTAG_1p_2m2t_32_b,
    ST3=all_expectedCountsTable$s315_SerC5_GGACTCCT_1p_2m2t_32_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_100p_2m2t <- round(data.frame(
    No1=all_expectedCountsTable$s308_SerB7_CTCTCTAC_ctrl_32_a, 
    No2=all_expectedCountsTable$s311_SerC1_TAAGGCGA_ctrl_32_b,
    No3=all_expectedCountsTable$s314_SerC4_TCCTGAGC_ctrl_32_c,
    ST1=all_expectedCountsTable$s310_SerB9_GCTACGCT_100p_2m2t_32_a,
    ST2=all_expectedCountsTable$s313_SerC3_AGGCAGAA_100p_2m2t_32_b,
    ST3=all_expectedCountsTable$s316_SerC6_TAGGCATG_100p_2m2t_32_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_sbt <- round(data.frame(
    No1=all_expectedCountsTable$s317_SerC7_CTCTCTAC_ctrl_33_a, 
    No2=all_expectedCountsTable$s320_SerC10_CGAGGCTG_ctrl_33_b,
    No3=all_expectedCountsTable$s323_SerD01_ctrl_33_c,
    ST1=all_expectedCountsTable$s318_SerC8_CAGAGAGG_1p_sbt_33_a,
    ST2=all_expectedCountsTable$s321_SerC11_AAGAGGCA_1p_sbt_33_b,
    ST3=all_expectedCountsTable$s324_SerD02_1p_sbt_33_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_100p_sbt <- round(data.frame(
    No1=all_expectedCountsTable$s317_SerC7_CTCTCTAC_ctrl_33_a, 
    No2=all_expectedCountsTable$s320_SerC10_CGAGGCTG_ctrl_33_b,
    No3=all_expectedCountsTable$s323_SerD01_ctrl_33_c,
    ST1=all_expectedCountsTable$s319_SerC9_GCTACGCT_100p_sbt_33_a,
    ST2=all_expectedCountsTable$s322_SerC12_GTAGAGGA_100p_sbt_33_b,
    ST3=all_expectedCountsTable$s325_SerD03_100p_sbt_33_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_ntmt <- round(data.frame(
    No1=all_expectedCountsTable$s326_SerD04_ctrl_34_a, 
    No2=all_expectedCountsTable$s329_SerD07_ctrl_34_b,
    No3=all_expectedCountsTable$s332_SerD10_ctrl_34_c,
    ST1=all_expectedCountsTable$s327_SerD05_1p_ntmt_34_a,
    ST2=all_expectedCountsTable$s330_SerD08_1p_ntmt_34_b,
    ST3=all_expectedCountsTable$s333_SerD11_1p_ntmt_34_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_p01_acetophenone <- round(data.frame(
    No1=all_expectedCountsTable$s335_SerE1_L001_R1_001_ctrl_35_a, 
    No2=all_expectedCountsTable$s338_SerE4_L001_R1_001_ctrl_35_b,
    No3=all_expectedCountsTable$s341_SerE7_L001_R1_001_ctrl_35_c,
    ST1=all_expectedCountsTable$s336_SerE2_L001_R1_001_p01_acetophenone_35_a,
    ST2=all_expectedCountsTable$s339_SerE5_L001_R1_001_p01_acetophenone_35_b,
    ST3=all_expectedCountsTable$s342_SerE8_L001_R1_001_p01_acetophenone_35_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_p1_acetophenone <- round(data.frame(
    No1=all_expectedCountsTable$s335_SerE1_L001_R1_001_ctrl_35_a, 
    No2=all_expectedCountsTable$s338_SerE4_L001_R1_001_ctrl_35_b,
    No3=all_expectedCountsTable$s341_SerE7_L001_R1_001_ctrl_35_c,
    ST1=all_expectedCountsTable$s337_SerE3_L001_R1_001_p1_acetophenone_35_a,
    ST2=all_expectedCountsTable$s340_SerE6_L001_R1_001_p1_acetophenone_35_b,
    ST3=all_expectedCountsTable$s343_SerE9_L001_R1_001_p1_acetophenone_35_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_p01_tmt <- round(data.frame(
    No1=all_expectedCountsTable$s350_SerG1_L001_R1_001_ctrl_36_a, 
    No2=all_expectedCountsTable$s354_SerG5_L001_R1_001_ctrl_36_b,
    No3=all_expectedCountsTable$s358_SerG9_L001_R1_001_ctrl_36_c,
    ST1=all_expectedCountsTable$s351_SerG2_L001_R1_001_p01_tmt_36_a,
    ST2=all_expectedCountsTable$s355_SerG6_L001_R1_001_p01_tmt_36_b,
    ST3=all_expectedCountsTable$s359_SerG10_L001_R1_001_p01_tmt_36_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_p1_tmt <- round(data.frame(
    No1=all_expectedCountsTable$s350_SerG1_L001_R1_001_ctrl_36_a, 
    No2=all_expectedCountsTable$s354_SerG5_L001_R1_001_ctrl_36_b,
    No3=all_expectedCountsTable$s358_SerG9_L001_R1_001_ctrl_36_c,
    ST1=all_expectedCountsTable$s352_SerG3_L001_R1_001_p1_tmt_36_a,
    ST2=all_expectedCountsTable$s356_SerG7_L001_R1_001_p1_tmt_36_b,
    ST3=all_expectedCountsTable$s360_SerG11_L001_R1_001_p1_tmt_36_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_10p_tmt <- round(data.frame(
    No1=all_expectedCountsTable$s350_SerG1_L001_R1_001_ctrl_36_a, 
    No2=all_expectedCountsTable$s354_SerG5_L001_R1_001_ctrl_36_b,
    No3=all_expectedCountsTable$s358_SerG9_L001_R1_001_ctrl_36_c,
    ST1=all_expectedCountsTable$s353_SerG4_L001_R1_001_10p_tmt_36_a,
    ST2=all_expectedCountsTable$s357_SerG8_L001_R1_001_10p_tmt_36_b,
    ST3=all_expectedCountsTable$s361_SerG12_L001_R1_001_10p_tmt_36_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_pLimonene <- round(data.frame(
    No1=all_expectedCountsTable$s362_SerH01_combine_ctrl_37_a, 
    No2=all_expectedCountsTable$s365_SerH04_combine_ctrl_37_b,
    No3=all_expectedCountsTable$s368_SerH07_combine_ctrl_37_c,
    ST1=all_expectedCountsTable$s363_SerH02_combine_1p_pLimonene_37_a,
    ST2=all_expectedCountsTable$s366_SerH05_combine_1p_pLimonene_37_b,
    ST3=all_expectedCountsTable$s369_SerH08_combine_1p_pLimonene_37_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_propionicAcid <- round(data.frame(
    No1=all_expectedCountsTable$s362_SerH01_combine_ctrl_37_a, 
    No2=all_expectedCountsTable$s365_SerH04_combine_ctrl_37_b,
    No3=all_expectedCountsTable$s368_SerH07_combine_ctrl_37_c,
    ST1=all_expectedCountsTable$s364_SerH03_combine_1p_propionicAcid_37_a,
    ST2=all_expectedCountsTable$s367_SerH06_combine_1p_propionicAcid_37_b,
    ST3=all_expectedCountsTable$s370_SerH09_combine_1p_propionicAcid_37_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_geranoil <- round(data.frame(
    No1=all_expectedCountsTable$s371_SerJ5_12072017_ctrl_38_a, 
    No2=all_expectedCountsTable$s375_SerJ9_12072017_ctrl_38_b,
    No3=all_expectedCountsTable$s379_SerF01_combine_ctrl_38_c,
    ST1=all_expectedCountsTable$s372_SerJ6_12072017_1p_geraniol_38_a,
    ST2=all_expectedCountsTable$s376_SerJ10_12072017_1p_geraniol_38_b,
    ST3=all_expectedCountsTable$s380_SerF02_combine_1p_geraniol_38_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_bIonone <- round(data.frame(
    No1=all_expectedCountsTable$s371_SerJ5_12072017_ctrl_38_a, 
    No2=all_expectedCountsTable$s375_SerJ9_12072017_ctrl_38_b,
    No3=all_expectedCountsTable$s379_SerF01_combine_ctrl_38_c,
    ST1=all_expectedCountsTable$s373_SerJ7_12072017_1p_bIonone_38_a,
    ST2=all_expectedCountsTable$s377_SerJ11_12072017_1p_bIonone_38_b,
    ST3=all_expectedCountsTable$s381_SerF03_combine_1p_bIonone_38_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_dimethylSulfide <- round(data.frame(
    No1=all_expectedCountsTable$s371_SerJ5_12072017_ctrl_38_a, 
    No2=all_expectedCountsTable$s375_SerJ9_12072017_ctrl_38_b,
    No3=all_expectedCountsTable$s379_SerF01_combine_ctrl_38_c,
    ST1=all_expectedCountsTable$s374_SerJ8_12072017_1p_dimethylsulfide_38_a,
    ST2=all_expectedCountsTable$s378_SerJ12_12072017_1p_dimethylsulfide_38_b,
    ST3=all_expectedCountsTable$s382_SerF04_combine_1p_dimethylsulfide_38_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_ethylTiglate <- round(data.frame(
    No1=all_expectedCountsTable$s383_SerF05_combine_ctrl_39_a, 
    No2=all_expectedCountsTable$s387_SerF09_combine_ctrl_39_b,
    No3=all_expectedCountsTable$s391_SerJ1_12072017_ctrl_39_c,
    ST1=all_expectedCountsTable$s384_SerF06_combine_1p_ethylTiglate_39_a,
    ST2=all_expectedCountsTable$s388_SerF10_combine_1p_ethylTiglate_39_b,
    ST3=all_expectedCountsTable$s392_SerJ2_12072017_1p_ethylTiglate_39_c,
    row.names=geneID_info$ensembl_gene_id))

pS6_countsTable_1p_heptanoicAcid <- round(data.frame(
    No1=all_expectedCountsTable$s383_SerF05_combine_ctrl_39_a, 
    No2=all_expectedCountsTable$s387_SerF09_combine_ctrl_39_b,
    No3=all_expectedCountsTable$s391_SerJ1_12072017_ctrl_39_c,
    ST1=all_expectedCountsTable$s386_SerF08_combine_1p_heptanoicAcid_39_a,
    ST2=all_expectedCountsTable$s390_SerF12_combine_1p_heptanoicAcid_39_b,
    ST3=all_expectedCountsTable$s394_SerJ4_12072017_1p_heptanoicAcid_39_c,
    row.names=geneID_info$ensembl_gene_id))


DiffExp <- function (targets, countsTable) {
    Treat <- factor(targets$Treatment);Subject <- factor(targets$Subject);design <- model.matrix(~Subject+Treat)
    e.litter <- DGEList(counts=countsTable)
    e.litter <- estimateGLMCommonDisp(e.litter,design)
    e.litter <- estimateGLMTrendedDisp(e.litter,design)
    e.litter <- estimateGLMTagwiseDisp(e.litter,design)
    fit <- glmFit(e.litter, design);lrt <- glmLRT(fit);diff <- topTags(lrt,n=dim(lrt)[1])$table
    result <- merge(diff,countsTable,by=0,sort=F)
    colnames(result)[1] <- "ensembl_gene_id"
    return(result)
}

DiffExp_compile <- function(targets, countsTable) {
    allGene_DE <- DiffExp(targets, countsTable)
    allGene_DE <- full_join(allGene_DE, geneID_info, by = "ensembl_gene_id")
    return(allGene_DE)
}


setwd("./../all_gene_de/")

pS6_DE_100mM_CPD <- DiffExp_compile(meta_ST3, pS6_countsTable_100mM_CPD)
write.csv(pS6_DE_100mM_CPD, "./pS6_DE_100mM_CPD.csv")
pS6_DE_10mM_MTMT <- DiffExp_compile(meta_ST3, pS6_countsTable_10mM_MTMT)
write.csv(pS6_DE_10mM_MTMT, "./pS6_DE_10mM_MTMT.csv")
pS6_DE_1p_Benzaldehyde <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_Benzaldehyde)
write.csv(pS6_DE_1p_Benzaldehyde, "./pS6_DE_1p_Benzaldehyde.csv")
pS6_DE_1p_Octanal <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_Octanal)
write.csv(pS6_DE_1p_Octanal, "./pS6_DE_1p_Octanal.csv")
pS6_DE_1p_pDihydrocarveol <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_pDihydrocarveol)
write.csv(pS6_DE_1p_pDihydrocarveol, "./pS6_DE_1p_pDihydrocarveol.csv")
pS6_DE_1p_nDihydrocarveol <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_nDihydrocarveol)
write.csv(pS6_DE_1p_nDihydrocarveol, "./pS6_DE_1p_nDihydrocarveol.csv")
pS6_DE_1p_citronellol <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_citronellol)
write.csv(pS6_DE_1p_citronellol, "./pS6_DE_1p_citronellol.csv")
pS6_DE_1p_guaiacol <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_guaiacol)
write.csv(pS6_DE_1p_guaiacol, "./pS6_DE_1p_guaiacol.csv")
pS6_DE_1p_pCarvone <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_pCarvone)
write.csv(pS6_DE_1p_pCarvone, "./pS6_DE_1p_pCarvone.csv")
pS6_DE_1p_nCarvone <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_nCarvone)
write.csv(pS6_DE_1p_nCarvone, "./pS6_DE_1p_nCarvone.csv")
pS6_DE_1p_isoamylAcetate <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_isoamylAcetate)
write.csv(pS6_DE_1p_isoamylAcetate, "./pS6_DE_1p_isoamylAcetate.csv")
pS6_DE_p1_isovalericAcid <- DiffExp_compile(meta_ST3, pS6_countsTable_p1_isovalericAcid)
write.csv(pS6_DE_p1_isovalericAcid, "./pS6_DE_p1_isovalericAcid.csv")
pS6_DE_3mM_androstenone <- DiffExp_compile(meta_ST3, pS6_countsTable_3mM_androstenone)
write.csv(pS6_DE_3mM_androstenone, "./pS6_DE_3mM_androstenone.csv")
pS6_DE_1p_bCaryophyllene <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_bCaryophyllene)
write.csv(pS6_DE_1p_bCaryophyllene, "./pS6_DE_1p_bCaryophyllene.csv")
pS6_DE_1p_pyridine <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_pyridine)
write.csv(pS6_DE_1p_pyridine, "./pS6_DE_1p_pyridine.csv")
pS6_DE_1p_methylSalicylate <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_methylSalicylate)
write.csv(pS6_DE_1p_methylSalicylate, "./pS6_DE_1p_methylSalicylate.csv")
pS6_DE_1p_2heptanone <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_2heptanone)
write.csv(pS6_DE_1p_2heptanone, "./pS6_DE_1p_2heptanone.csv")
pS6_DE_1p_linalool <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_linalool)
write.csv(pS6_DE_1p_linalool, "./pS6_DE_1p_linalool.csv")
pS6_DE_1p_ethylButyrate <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_ethylButyrate)
write.csv(pS6_DE_1p_ethylButyrate, "./pS6_DE_1p_ethylButyrate.csv")
pS6_DE_1M_pMenthol <- DiffExp_compile(meta_ST3, pS6_countsTable_1M_pMenthol)
write.csv(pS6_DE_1M_pMenthol, "./pS6_DE_1M_pMenthol.csv")
pS6_DE_1M_nMenthol <- DiffExp_compile(meta_ST3, pS6_countsTable_1M_nMenthol)
write.csv(pS6_DE_1M_nMenthol, "./pS6_DE_1M_nMenthol.csv")
pS6_DE_100p_ebFarnesene <- DiffExp_compile(meta_ST3, pS6_countsTable_100p_ebFarnesene)
write.csv(pS6_DE_100p_ebFarnesene, "./pS6_DE_100p_ebFarnesene.csv")
pS6_DE_1p_25dmp <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_25dmp)
write.csv(pS6_DE_1p_25dmp, "./pS6_DE_1p_25dmp.csv")
pS6_DE_100uM_MTMT <- DiffExp_compile(meta_ST3, pS6_countsTable_100uM_MTMT)
write.csv(pS6_DE_100uM_MTMT, "./pS6_DE_100uM_MTMT.csv")
pS6_DE_10p_p2octanol <- DiffExp_compile(meta_ST3, pS6_countsTable_10p_p2octanol)
write.csv(pS6_DE_10p_p2octanol, "./pS6_DE_10p_p2octanol.csv")
pS6_DE_10p_n2octanol <- DiffExp_compile(meta_ST3, pS6_countsTable_10p_n2octanol)
write.csv(pS6_DE_10p_n2octanol, "./pS6_DE_10p_n2octanol.csv")
pS6_DE_77p_dhb <- DiffExp_compile(meta_ST3, pS6_countsTable_77p_dhb)
write.csv(pS6_DE_77p_dhb, "./pS6_DE_77p_dhb.csv")
pS6_DE_p01_sbt <- DiffExp_compile(meta_ST3, pS6_countsTable_p01_sbt)
write.csv(pS6_DE_p01_sbt, "./pS6_DE_p01_sbt.csv")
pS6_DE_1p_aPinene <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_aPinene)
write.csv(pS6_DE_1p_aPinene, "./pS6_DE_1p_aPinene.csv")
pS6_DE_1p_bDamascone <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_bDamascone)
write.csv(pS6_DE_1p_bDamascone, "./pS6_DE_1p_bDamascone.csv")
pS6_DE_1p_2m2p <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_2m2p)
write.csv(pS6_DE_1p_2m2p, "./pS6_DE_1p_2m2p.csv")
pS6_DE_1p_transCinnamaldehyde <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_transCinnamaldehyde)
write.csv(pS6_DE_1p_transCinnamaldehyde, "./pS6_DE_1p_transCinnamaldehyde.csv")
pS6_DE_1p_hexylTiglate <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_hexylTiglate)
write.csv(pS6_DE_1p_hexylTiglate, "./pS6_DE_1p_hexylTiglate.csv")
pS6_DE_1p_2e3mp <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_2e3mp)
write.csv(pS6_DE_1p_2e3mp, "./pS6_DE_1p_2e3mp.csv")
pS6_DE_1p_isopropylTiglate <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_isopropylTiglate)
write.csv(pS6_DE_1p_isopropylTiglate, "./pS6_DE_1p_isopropylTiglate.csv")
pS6_DE_1p_2hac <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_2hac)
write.csv(pS6_DE_1p_2hac, "./pS6_DE_1p_2hac.csv")
pS6_DE_1p_heptanal <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_heptanal)
write.csv(pS6_DE_1p_heptanal, "./pS6_DE_1p_heptanal.csv")
pS6_DE_1p_4methylAC <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_4methylAC)
write.csv(pS6_DE_1p_4methylAC, "./pS6_DE_1p_4methylAC.csv")
pS6_DE_1p_3methyl1butanethiol <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_3methyl1butanethiol)
write.csv(pS6_DE_1p_3methyl1butanethiol, "./pS6_DE_1p_3methyl1butanethiol.csv")
pS6_DE_1p_paraCresol <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_paraCresol)
write.csv(pS6_DE_1p_paraCresol, "./pS6_DE_1p_paraCresol.csv")
pS6_DE_1p_dimethyltrisulfide <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_dimethyltrisulfide)
write.csv(pS6_DE_1p_dimethyltrisulfide, "./pS6_DE_1p_dimethyltrisulfide.csv")
pS6_DE_1p_butyricAcid <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_butyricAcid)
write.csv(pS6_DE_1p_butyricAcid, "./pS6_DE_1p_butyricAcid.csv")
pS6_DE_1p_indole <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_indole)
write.csv(pS6_DE_1p_indole, "./pS6_DE_1p_indole.csv")
pS6_DE_1p_tbm <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_tbm)
write.csv(pS6_DE_1p_tbm, "./pS6_DE_1p_tbm.csv")
pS6_DE_1p_diacetyl <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_diacetyl)
write.csv(pS6_DE_1p_diacetyl, "./pS6_DE_1p_diacetyl.csv")
pS6_DE_1p_2hexanone <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_2hexanone)
write.csv(pS6_DE_1p_2hexanone, "./pS6_DE_1p_2hexanone.csv")
pS6_DE_1p_2phenylAlcohol <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_2phenylAlcohol)
write.csv(pS6_DE_1p_2phenylAlcohol, "./pS6_DE_1p_2phenylAlcohol.csv")
pS6_DE_p01_e2butene1thiol <- DiffExp_compile(meta_ST3, pS6_countsTable_p01_e2butene1thiol)
write.csv(pS6_DE_p01_e2butene1thiol, "./pS6_DE_p01_e2butene1thiol.csv")
pS6_DE_p01_cyclopentanethiol <- DiffExp_compile(meta_ST3, pS6_countsTable_p01_cyclopentanethiol)
write.csv(pS6_DE_p01_cyclopentanethiol, "./pS6_DE_p01_cyclopentanethiol.csv")
pS6_DE_1p_citral <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_citral)
write.csv(pS6_DE_1p_citral, "./pS6_DE_p01_citral.csv")
pS6_DE_1p_anisaldehyde <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_anisaldehyde)
write.csv(pS6_DE_1p_anisaldehyde, "./pS6_DE_p01_anisaldehyde.csv")
pS6_DE_1p_acetophenone <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_acetophenone)
write.csv(pS6_DE_1p_acetophenone, "./pS6_DE_1p_acetophenone.csv")
pS6_DE_100p_acetophenone <- DiffExp_compile(meta_ST3, pS6_countsTable_100p_acetophenone)
write.csv(pS6_DE_100p_acetophenone, "./pS6_DE_100p_acetophenone.csv")
pS6_DE_1p_tmt <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_tmt)
write.csv(pS6_DE_1p_tmt, "./pS6_DE_1p_tmt.csv")
pS6_DE_100p_tmt <- DiffExp_compile(meta_ST3, pS6_countsTable_100p_tmt)
write.csv(pS6_DE_100p_tmt, "./pS6_DE_100p_tmt.csv")
pS6_DE_1p_2m2t <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_2m2t)
write.csv(pS6_DE_1p_2m2t, "./pS6_DE_1p_2m2t.csv")
pS6_DE_100p_2m2t <- DiffExp_compile(meta_ST3, pS6_countsTable_100p_2m2t)
write.csv(pS6_DE_100p_2m2t, "./pS6_DE_100p_2m2t.csv")
pS6_DE_1p_sbt <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_sbt)
write.csv(pS6_DE_1p_sbt, "./pS6_DE_1p_sbt.csv")
pS6_DE_100p_sbt <- DiffExp_compile(meta_ST3, pS6_countsTable_100p_sbt)
write.csv(pS6_DE_100p_sbt, "./pS6_DE_100p_sbt.csv")
pS6_DE_1p_ntmt <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_ntmt)
write.csv(pS6_DE_1p_ntmt, "./pS6_DE_1p_ntmt.csv")
pS6_DE_p01_acetophenone <- DiffExp_compile(meta_ST3, pS6_countsTable_p01_acetophenone)
write.csv(pS6_DE_p01_acetophenone, "./pS6_DE_p01_acetophenone.csv")
pS6_DE_p1_acetophenone <- DiffExp_compile(meta_ST3, pS6_countsTable_p1_acetophenone)
write.csv(pS6_DE_p1_acetophenone, "./pS6_DE_p1_acetophenone.csv")
pS6_DE_p01_tmt <- DiffExp_compile(meta_ST3, pS6_countsTable_p01_tmt)
write.csv(pS6_DE_p01_tmt, "./pS6_DE_p01_tmt.csv")
pS6_DE_p1_tmt <- DiffExp_compile(meta_ST3, pS6_countsTable_p1_tmt)
write.csv(pS6_DE_p1_tmt, "./pS6_DE_p1_tmt.csv")
pS6_DE_10p_tmt <- DiffExp_compile(meta_ST3, pS6_countsTable_10p_tmt)
write.csv(pS6_DE_10p_tmt, "./pS6_DE_10p_tmt.csv")
pS6_DE_1p_pLimonene <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_pLimonene)
write.csv(pS6_DE_1p_pLimonene, "./pS6_DE_1p_pLimonene.csv")
pS6_DE_1p_propionicAcid <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_propionicAcid)
write.csv(pS6_DE_1p_propionicAcid, "./pS6_DE_1p_propionicAcid.csv")
pS6_DE_1p_geranoil <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_geranoil)
write.csv(pS6_DE_1p_geranoil, "./pS6_DE_1p_geranoil.csv")
pS6_DE_1p_bIonone <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_bIonone)
write.csv(pS6_DE_1p_bIonone, "./pS6_DE_1p_bIonone.csv")
pS6_DE_1p_dimethylSulfide <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_dimethylSulfide)
write.csv(pS6_DE_1p_dimethylSulfide, "./pS6_DE_1p_dimethylSulfide.csv")
pS6_DE_1p_ethylTiglate <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_ethylTiglate)
write.csv(pS6_DE_1p_ethylTiglate, "./pS6_DE_1p_ethylTiglate.csv")
pS6_DE_1p_heptanoicAcid <- DiffExp_compile(meta_ST3, pS6_countsTable_1p_heptanoicAcid)
write.csv(pS6_DE_1p_heptanoicAcid, "./pS6_DE_1p_heptanoicAcid.csv")


aligned_olfr <- read.delim("./../aligned_olfr_fasta.csv", header=F, sep=',')
colnames(aligned_olfr) <- c("olfr","sequence")

filenames <- list.files("./", pattern="*.csv", full.names=FALSE)

allGene_df <- lapply(filenames, read.delim, sep=',',header=T, row.names = 1)
names(allGene_df) <- filenames

subset_olfr <- function(df){
    df <- df[grep("^Olfr", df$symbol),]
    df <- df[-grep("-ps", df$symbol),]
    df <- df[df$symbol %in% aligned_olfr$olfr,]
    df$FDR <- p.adjust(df$PValue, method='fdr')
    return(df)
}

olfrOnly_df <- lapply(allGene_df, subset_olfr)

setwd("./../olfr_de/")

for (i in seq(1,length(olfrOnly_df),1)){
    write.csv(olfrOnly_df[[i]], filenames[i])
}
