==> 4_Compile_descriptors_and_responses.ipynb <==
import pandas as pd
import numpy as np
import scipy

from scipy.spatial.distance import pdist
from scipy import stats

import statistics
import itertools
import requests
import os
import json
from time import sleep
from subprocess import Popen

print(pd.__version__)
print(np.__version__)
print(scipy.__version__)
print(requests.__version__)


untested_cids = pd.read_csv("./cid_info/untested_CID_list.csv", index_col = 0)
untested_cids = untested_cids.iloc[:,0]
untested_cids = list(map(str, untested_cids))
print(len(untested_cids))

tested_filename_cid = pd.read_csv("./cid_info/filename_cid.csv")
tested_filename_cid = tested_filename_cid.iloc[:,1]
tested_filename_cid = list(map(str, tested_filename_cid))

tested_cids = []
for cid in tested_filename_cid:
    if cid not in tested_cids:
        tested_cids.append(cid)
 
print(len(tested_cids))

untested_sdf = "./untested_odor_sdf/"
tested_sdf = "./tested_odor_sdf/"

untested_desc = "./untested_odor_desc/"
tested_desc = "./tested_odor_desc/"

# Download SDF files from PubChem

Uncomment to download

#for cid in tested_cids:
#    download_url = f'https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/CID/{cid}/record/SDF/?record_type=3d&response_type=save'
#    download_file = requests.get(download_url, allow_redirects=True)
#    target_location = os.path.join(tested_sdf, f'{cid}.sdf')
#    open(target_location, 'wb').write(download_file.content)
#    sleep(0.5) 

#for cid in untested_cids:
#    download_url = f'https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/CID/{cid}/record/SDF/?record_type=3d&response_type=save'
#    download_file = requests.get(download_url, allow_redirects=True)
#    target_location = os.path.join(untested_sdf, f'{cid}.sdf')
#    open(target_location, 'wb').write(download_file.content)
#    sleep(0.5)


# Run AlvaDesc to calculate descriptors

Uncomment to run

alvadesc = "../alvaDesc_Linux_64_v2_0_10/usr/bin/alvaDescCLI"

#for sdf in os.listdir(tested_sdf):
    #in_path = os.path.join(tested_sdf, sdf)
    #out_path = os.path.join(tested_desc, sdf.split('.')[0]+".txt")
    #Popen([alvadesc, "--input="+in_path,
    #       "--descriptors=ALL",
    #       "--output="+out_path,
    #       "--labels"])
    #sleep(0.5)

#for sdf in os.listdir(untested_sdf):
#    in_path = os.path.join(untested_sdf, sdf)
#    out_path = os.path.join(untested_desc, sdf.split('.')[0]+".txt")
#    Popen([alvadesc, "--input="+in_path,
#           "--descriptors=ALL",
#           "--output="+out_path,
#           "--labels"])
#    sleep(0.5)

# Compile descriptors

tested_compiled_out = "./compiled_desc_resp/compiled_tested_desc.txt"
untested_compiled_out = "./compiled_desc_resp/compiled_untested_desc.txt"

#for file in os.listdir(tested_desc):
#    file_path = os.path.join(tested_desc, file)
#    with open(file_path, 'r') as in_file:
#        next(in_file)
#        second_line = in_file.readline()
#    in_file.close()
#    with open(tested_compiled_out, 'a') as out_file:
#        out_file.write(second_line)
#    out_file.close()

#for file in os.listdir(untested_desc):
#    file_path = os.path.join(untested_desc, file)
#    with open(file_path, 'r') as in_file:
#        next(in_file)
#        second_line = in_file.readline()
#    in_file.close()
#    with open(untested_compiled_out, 'a') as out_file:
#        out_file.write(second_line)
#    out_file.close()

#with open(file_path, 'r') as in_file:
#    descs = in_file.readline()
#in_file.close()
#
#with open(tested_compiled_out, 'a') as out_file:
#    out_file.write(descs)
#out_file.close()
#
#with open(untested_compiled_out, 'a') as out_file:
#    out_file.write(descs)
#out_file.close()

#tested_comp_descs = pd.read_csv(tested_compiled_out, sep='\t', header=None, dtype = object)
#untested_comp_descs = pd.read_csv(untested_compiled_out, sep='\t', header=None, dtype = object)
#
#tested_comp_descs.columns = tested_comp_descs.iloc[tested_comp_descs.shape[0]-1,:]
#untested_comp_descs.columns = untested_comp_descs.iloc[untested_comp_descs.shape[0]-1,:]
#
#tested_comp_descs = tested_comp_descs.drop(axis = 0, index = tested_comp_descs.shape[0]-1)
#untested_comp_descs = untested_comp_descs.drop(axis = 0, index = untested_comp_descs.shape[0]-1)
#
#tested_comp_descs = tested_comp_descs.drop(axis = 1, columns = ["No."])
#untested_comp_descs = untested_comp_descs.drop(axis = 1, columns = ["No."])
#
#tested_comp_descs.to_csv("./compiled_desc_resp/compiled_tested_desc.csv")
#untested_comp_descs.to_csv("./compiled_desc_resp/compiled_untested_desc.csv")
#
#tested_comp_descs = pd.read_csv("./compiled_desc_resp/compiled_tested_desc.csv", index_col = 1, dtype = object)
#untested_comp_descs = pd.read_csv("./compiled_desc_resp/compiled_untested_desc.csv", index_col = 1, dtype = object)
#
#tested_comp_descs = tested_comp_descs.drop(axis = 1, columns = ["Unnamed: 0"])
#untested_comp_descs = untested_comp_descs.drop(axis = 1, columns = ["Unnamed: 0"])
#
#tested_comp_descs.index.name = None
#untested_comp_descs.index.name = None
#
#tested_comp_descs.to_csv("./compiled_desc_resp/compiled_tested_desc.csv")
#untested_comp_descs.to_csv("./compiled_desc_resp/compiled_untested_desc.csv")

# Preprocess descriptors

tested_comp_descs = pd.read_csv("./compiled_desc_resp/compiled_tested_desc.csv", index_col = 0, dtype = object)
untested_comp_descs = pd.read_csv("./compiled_desc_resp/compiled_untested_desc.csv", index_col = 0, dtype = object)

tested_comp_descs.index = tested_comp_descs.index.map(str)
untested_comp_descs.index = untested_comp_descs.index.map(str)

untested_comp_descs['MW'] = pd.to_numeric(untested_comp_descs['MW'])
#Molecular weight cutoffs from https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4977894/
untested_comp_descs = untested_comp_descs[~(untested_comp_descs['MW'] > 402.55)]
untested_comp_descs = untested_comp_descs[~(untested_comp_descs['MW'] < 18.02)]
print(untested_comp_descs.shape)
#(4671, 5666)

#Recombine tested+untested after removing unwanted odors from untested odors
all_comp_descs = pd.concat([tested_comp_descs, untested_comp_descs])
#Drop features that have 'na' values from AlvaDesc
all_comp_descs = all_comp_descs.loc[:,~all_comp_descs.isin(["na"]).any()]
print(all_comp_descs.shape)
#(4732, 4302)
all_comp_descs = all_comp_descs.apply(pd.to_numeric)
all_comp_descs = all_comp_descs.loc[:,~(all_comp_descs.var()==0)]
print(all_comp_descs.shape)
#(4732, 2660)

#Subset the kept features in the tested_desc df
tested_comp_descs = tested_comp_descs.loc[:,tested_comp_descs.columns.isin(all_comp_descs.columns)]
tested_comp_descs = tested_comp_descs.apply(pd.to_numeric)
tested_comp_descs = tested_comp_descs.loc[:,~(tested_comp_descs.var()==0)]
print(tested_comp_descs.shape)
#(61, 1841)
all_comp_descs = all_comp_descs.loc[:,all_comp_descs.columns.isin(tested_comp_descs.columns)]
print(all_comp_descs.shape)
#(4732, 1841)

tested_comp_descs.to_csv("./compiled_desc_resp/filtered_tested_desc.csv")
all_comp_descs.to_csv("./compiled_desc_resp/filtered_all_desc.csv")

tested_comp_descs = pd.read_csv("./compiled_desc_resp/filtered_tested_desc.csv", index_col = 0)
tested_comp_descs = tested_comp_descs.sort_index(axis = 0)
all_comp_descs = pd.read_csv("./compiled_desc_resp/filtered_all_desc.csv", index_col = 0)

filename_cid = pd.read_csv("./cid_info/filename_cid.csv", index_col = 0)
filename_cid = filename_cid.to_dict(orient='index')

ps6_dir = "./olfr_de"

odor_test_count = {}

for base_name in filename_cid:
    split_name = base_name.split('_')[2:]
    odor_name = split_name[1].split('.')[0]
    if odor_name not in odor_test_count:
        odor_test_count[odor_name] = 0
    odor_test_count[odor_name] += 1
    
    multi_conc_tested = {}

for ps6ip_file in os.listdir(ps6_dir):
    conc_odor = ps6ip_file.split('_')[2:]
    conc = conc_odor[0]
    odor_name = conc_odor[1].split('.')[0]
    if odor_test_count[odor_name] > 1:
        ps6ip_file = os.path.join(ps6_dir, ps6ip_file)
        df = pd.read_csv(ps6ip_file, index_col = 0)
        if odor_name not in multi_conc_tested:
            multi_conc_tested[odor_name] = {}
        multi_conc_tested[odor_name][conc] = df
        
multi_conc_activation = {}

for odor in multi_conc_tested:
    if odor not in multi_conc_activation:
        multi_conc_activation[odor] = {}
    for conc in multi_conc_tested[odor]:
        df = multi_conc_tested[odor][conc]
        sig_or_count = df[(df.logFC > 0) & (df.FDR < 0.05)].shape[0]
        multi_conc_activation[odor][conc] = sig_or_count
       

tested_resp = {}
sigOR_dict = {}

for odor in odor_test_count:
    #Pick out concentration for odors tested at multiple concentrations
    if odor_test_count[odor] > 1:
        fewest_or_conc = min(multi_conc_activation[odor], key=multi_conc_activation[odor].get)
        filename = "pS6_DE_"+fewest_or_conc+"_"+odor+".csv"
    else:
    #Rest which are tested at a single concentration
        for base_file in os.listdir(ps6_dir):
            odor_name = base_file.split('_')[3].split('.')[0]
            if odor == odor_name:
                filename = base_file
    file_path = os.path.join(ps6_dir, filename)
    cid = str(filename_cid[filename]['cid'])
    df = pd.read_csv(file_path, index_col = 0)
    df = df.loc[:,['symbol','logFC','FDR']]
    df = df.sort_values(by=['symbol'])
    df = df.reset_index(drop=True)
    #Set criteria for determining significant response
    sigOR_count = df[(df.logFC > 0) & (df.FDR < 0.05)].shape[0]
    if sigOR_count == 0:
        #Uncomment following line if running for first time
        tested_comp_descs = tested_comp_descs.drop([int(cid)])
        print("No responsive ORs:", cid)
    sigOR_dict[odor] = sigOR_count
    df0 = df.copy(deep=True)
    #Set criteria for determining non-significant response
    df0.loc[(df0.logFC <= 0) | (df0.FDR >= 0.05),'logFC'] = 0
    df.columns = df.columns+"_"+cid
    df0.columns = df0.columns+"_"+cid
    if cid not in tested_resp:
        tested_resp[cid] = {}
    tested_resp[cid]['orig'] = df
    tested_resp[cid]['zero'] = df0
#Determine mean & stdev of number of responsive receptors per odor
resp_dist = []
for odor in sigOR_dict:
    resp_dist.append(sigOR_dict[odor])
resp_cutoff = statistics.mean(resp_dist)+(5*statistics.stdev(resp_dist))
#Don't consider odors with activated OR # > 5 stdev (resp_cutoff)
for odor in sigOR_dict:
    if sigOR_dict[odor] > resp_cutoff:
        for base_file in os.listdir(ps6_dir):
            odor_name = base_file.split('_')[3].split('.')[0]
            if odor == odor_name:
                filename = base_file
                cid = filename_cid[filename]['cid']
                #Uncomment following line if running for first time
                tested_comp_descs = tested_comp_descs.drop([int(cid)])
                print("OR response count > 5 stdev:", cid)
trueResponse_df = pd.DataFrame()
zerodResponse_df = pd.DataFrame()
#Compile dictionary into logFC df
for cid in tested_comp_descs.index.values:
    df = tested_resp[str(cid)]['orig']
    df0 = tested_resp[str(cid)]['zero']
    trueResponse_df = pd.concat([trueResponse_df, df.iloc[:,1]], axis = 1)
    zerodResponse_df = pd.concat([zerodResponse_df, df0.iloc[:,1]], axis = 1)
#Change index to Olfr IDs
trueResponse_df.index = df.iloc[:,0].values
zerodResponse_df.index = df.iloc[:,0].values
#Drop non-responding receptors
zerodResponse_df = zerodResponse_df[~(zerodResponse_df.var(axis = 1) == 0)]
#Convert to dictionary
trueResponse_dict = trueResponse_df.to_dict()
zerodResponse_dict = zerodResponse_df.to_dict()
#Restore responses to de-orphaned Olfrs
for cid in zerodResponse_dict:
    for olfr in zerodResponse_dict[cid]:
        if zerodResponse_dict[cid][olfr] == 0:
            zerodResponse_dict[cid][olfr] = trueResponse_dict[cid][olfr]
logFC_restored = pd.DataFrame.from_dict(zerodResponse_dict)
logFC_restored = logFC_restored.transpose()
indices_nm = logFC_restored.index.values.tolist()
indices_nm = [i.split('_', 1)[1] for i in indices_nm]
indices_nm = list(map(int, indices_nm))
logFC_restored.index = indices_nm
#Sort index so that it matches descriptors
logFC_restored = logFC_restored.sort_index(axis = 0)
#Save logFC restored
#logFC_restored.to_csv("./compiled_desc_resp/compiled_odor_sigResp_wide.csv")
#Update descs
tested_comp_descs = tested_comp_descs.loc[:,-(tested_comp_descs.var() == 0)]
all_comp_descs = all_comp_descs.loc[:,all_comp_descs.columns.isin(tested_comp_descs.columns)]
#tested_comp_descs.to_csv("./compiled_desc_resp/filtered_tested_desc.csv")
#all_comp_descs.to_csv("./compiled_desc_resp/filtered_all_desc.csv")

# Olfr * Feature analysis

tested_odor_desc = pd.read_csv("./compiled_desc_resp/filtered_tested_desc.csv", index_col = 0)
tested_odor_resp = pd.read_csv("./compiled_desc_resp/compiled_odor_sigResp_wide.csv", index_col = 0)


#Function to zero center and scale data
def normalize(x):
    return (x - x.mean())/(x - x.mean()).std()

#Center-scale responses and descriptors by odors
norm_tested_desc = tested_odor_desc.transform(normalize, axis = 0)
norm_tested_resp = tested_odor_resp.transform(normalize, axis = 0)
#Compute Pearson correlation values
nOdors = norm_tested_resp.shape[0]
olfr_feat_pc = (norm_tested_resp.transpose() @ norm_tested_desc)/nOdors
#Compute t-statistic
num_val = olfr_feat_pc*np.sqrt(nOdors-2)
denom_val = np.sqrt(1 - (olfr_feat_pc**2))
olfr_feat_tStat = num_val/denom_val
#Calculate p-values
olfr_feat_pVal = stats.t.sf(abs(olfr_feat_tStat), nOdors-1)*2
olfr_feat_pVal = pd.DataFrame(olfr_feat_pVal)
olfr_feat_pVal.columns = norm_tested_desc.columns
olfr_feat_pVal = olfr_feat_pVal.set_index(norm_tested_resp.transpose().index)
#Melt data
olfr_feat_pc = olfr_feat_pc.reset_index().melt('index')
olfr_feat_pc.columns = ['Olfr','feature','pcorr']
olfr_feat_tStat = olfr_feat_tStat.reset_index().melt('index')
olfr_feat_tStat.columns = ['Olfr','feature','tStat']
olfr_feat_pVal = olfr_feat_pVal.reset_index().melt('index')
olfr_feat_pVal.columns = ['Olfr','feature','pVal']
#Compile data into a single df
olfr_feat_stats = pd.concat([olfr_feat_pc,
                             olfr_feat_tStat.loc[:,"tStat"],
                             olfr_feat_pVal.loc[:,"pVal"]],
                             axis = 1)
#Calculate distances
norm_tested_resp_dist_euc = pd.DataFrame(pdist(norm_tested_resp, 'euclidean'))
norm_tested_desc_dist_euc = pd.DataFrame(pdist(norm_tested_desc, 'euclidean'))
norm_tested_resp_dist_corr = pd.DataFrame(pdist(norm_tested_resp, 'correlation'))
norm_tested_desc_dist_corr = pd.DataFrame(pdist(norm_tested_desc, 'correlation'))
norm_tested_resp_dist_cos = pd.DataFrame(pdist(norm_tested_resp, 'cosine'))
norm_tested_desc_dist_cos = pd.DataFrame(pdist(norm_tested_desc, 'cosine'))
#Combine distances
odor_combns = pd.DataFrame(itertools.combinations(norm_tested_resp.index.values, 2))
odor_combns = pd.concat([odor_combns, norm_tested_resp_dist_euc,
                         norm_tested_desc_dist_euc,norm_tested_resp_dist_corr,
                         norm_tested_desc_dist_corr,norm_tested_resp_dist_cos,
                         norm_tested_desc_dist_cos], axis = 1)
odor_combns.columns = ['odor1','odor2',
                       'response_euc_distance','feature_euc_distance',
                       'response_corr_distance','feature_corr_distance',
                       'response_cos_distance','feature_cos_distance']
#olfr_feat_stats.to_csv("./compiled_desc_resp/odor_response_stats.csv")
#odor_combns.to_csv("./compiled_desc_resp/odor_resp_distances.csv")

compiled_ps6_data = pd.DataFrame()

for cid in tested_odor_desc.index.values:
    df = tested_resp[str(cid)]['orig']
    df.columns = ['symbol','logFC','FDR']
    df['cid'] = cid
    df = df.loc[df['symbol'].isin(tested_odor_resp.columns.values),:]
    df = df.reset_index(drop=True)
    compiled_ps6_data = pd.concat([compiled_ps6_data, df], axis = 0)
compiled_ps6_data = compiled_ps6_data.reset_index(drop=True)
#compiled_ps6_data.to_csv("./compiled_desc_resp/compiled_odor_sigResp_long.csv")

odorinfo = pd.read_csv("./cid_info/cid_odornames.csv", index_col = 0)
#Calculate pairwise odor similarity and shared # of responding receptors
t_tested_odor_resp = tested_odor_resp.transpose()
t_tested_odor_resp = t_tested_odor_resp.transform(normalize)
cor_matrix = t_tested_odor_resp.corr()
cor_matrix = 1 - cor_matrix
cor_matrix = cor_matrix.where(np.triu(np.ones(cor_matrix.shape)).astype(bool))
cor_matrix = cor_matrix.stack().reset_index()
cor_matrix.columns = ['odor1','odor2','pearson_corr_dist']
cor_matrix = cor_matrix[cor_matrix['pearson_corr_dist'] != 0]
resp_list = []
o1_names_list = []
o2_names_list = []
for i in range(cor_matrix.shape[0]):
    o1 = str(int(cor_matrix.iloc[i,:]['odor1']))
    o2 = str(int(cor_matrix.iloc[i,:]['odor2']))
    o1_df = tested_resp[o1]['orig']
    o2_df = tested_resp[o2]['orig']
    o1_rec = o1_df[(o1_df.iloc[:,1] > 0) & (o1_df.iloc[:,2] < 0.05)].iloc[:,0].values.tolist()
    o2_rec = o2_df[(o2_df.iloc[:,1] > 0) & (o2_df.iloc[:,2] < 0.05)].iloc[:,0].values.tolist()
    o1_name = odorinfo[odorinfo['cid'] == int(o1)]['odor_list'].values[0]
    o2_name = odorinfo[odorinfo['cid'] == int(o2)]['odor_list'].values[0]
    o1_names_list.append(o1_name)
    o2_names_list.append(o2_name)
    o1_as_set = set(o1_rec)
    intersection = o1_as_set.intersection(o2_rec)
    intersection_as_list = list(intersection)
    resp_list.append(len(intersection_as_list))
cor_matrix['o1_name'] = o1_names_list
cor_matrix['o2_name'] = o2_names_list
cor_matrix['shared_receptor_count'] = resp_list
#cor_matrix.to_csv("./compiled_desc_resp/odor_response_similarities.csv")

# Compile full pS6-IP-Seq dataset

#olfr_de_dir = "./olfr_de/"
#
#full_compiled_ps6 = pd.DataFrame()
#
#for file in os.listdir(olfr_de_dir):
#    conc = file.split('_')[2]
#    odor = file.split('_')[3].split('.')[0]
#    conc_odor = str(conc+"_"+odor)
#    file_path = os.path.join(olfr_de_dir, file)
#    read_df = pd.read_csv(file_path, index_col = 0)
#    read_df = read_df.iloc[:,[1,5,12]]
#    read_df = read_df.sort_values('symbol')
#    read_df.columns = read_df.add_prefix(conc_odor+'_').columns.tolist()
#    read_df = read_df.reset_index(drop=True)
#    full_compiled_ps6 = pd.concat([full_compiled_ps6, read_df], axis = 1)
#    
#sig_full_ps6 = pd.DataFrame()
#for i in range(full_compiled_ps6.shape[0]):
#    for j in range(0,len(full_compiled_ps6.iloc[i,:]),3):
#        if (full_compiled_ps6.iloc[i,:][j] > 0) & (full_compiled_ps6.iloc[i,:][j+1] < 0.05):
#            sig_full_ps6 = pd.concat([sig_full_ps6, full_compiled_ps6.iloc[i,:]], axis = 1)
#            break
#            
#sig_full_ps6 = sig_full_ps6.transpose()
#sig_full_ps6.index = sig_full_ps6.iloc[:,2]
#sig_full_ps6.index.name = None
#
#drop_cols = []
#for i in range(0,sig_full_ps6.shape[1],3):
#    drop_cols.append(i+1)
#    drop_cols.append(i+2)
#sig_full_ps6.drop(sig_full_ps6.columns[drop_cols],axis=1,inplace=True)
#
#sig_full_ps6.to_csv("./compiled_desc_resp/compiled_allOdor_sigResp_wide.csv")
sig_full_ps6 = pd.read_csv("./compiled_desc_resp/compiled_allOdor_sigResp_wide.csv", index_col = 0)
