==> 15_Make_supp_tables.ipynb <==
import pandas as pd
import numpy as np
import itertools
import os

import sklearn
import scipy

from scipy.spatial.distance import pdist
from sklearn.metrics import roc_curve, auc

print(pd.__version__)
#1.3.3
print(np.__version__)
#1.20.3
print(sklearn.__version__)
#0.24.2
print(scipy.__version__)
#1.7.1

# Supp table 1

#cid_conv = pd.read_csv('./cid_info/filename_cid.csv')
#compiled_ps6 = pd.DataFrame()
#
#for file in os.listdir("./olfr_de/"):
#    file_path = os.path.join("./olfr_de",file)
#    ps6_dat = pd.read_csv(file_path, index_col = 0)
#    ps6_dat = ps6_dat.loc[:,['symbol','logFC','FDR']]
#    conc = file.split('_')[2]
#    odor = file.split('_')[3]
#    odor = odor.split('.')[0]
#    cid = cid_conv[cid_conv['filename'] == file]['cid'].values[0]
#    ps6_dat['conc'] = conc
#    ps6_dat['odor'] = odor
#    ps6_dat['cid'] = cid
#    ps6_dat = ps6_dat.reset_index(drop=True)
#    compiled_ps6 = pd.concat([compiled_ps6, ps6_dat])
#    
#compiled_ps6 = compiled_ps6.sort_values(['FDR', 'logFC'])
#compiled_ps6.columns = ['Olfr', 'logFC', 'FDR', 'concentration', 'odor', 'cid']
#compiled_ps6 = compiled_ps6.reset_index(drop = True)
#compiled_ps6.to_csv("./supp_tables/Supplementary_table_1_pS6_data.csv")
s1_tbl = pd.read_csv("./supp_tables/Supplementary_table_1_pS6_data.csv", index_col = 0)

# Supp table 2

#opt_descs = pd.read_csv("./compiled_desc_resp/well_predicted_desc.csv", index_col = 0)
#alva = pd.read_csv("./compiled_desc_resp/AlvaDesc_descriptors.csv", index_col = 0)
#
#
#s1_tbl = pd.DataFrame()
#for desc in opt_descs.columns:
#    long = alva[alva['Abbrev'] == desc]['Long'].values[0]
#    cat = alva[alva['Abbrev'] == desc]['Category'].values[0]
#    temp = pd.DataFrame([desc,long, cat]).transpose()
#    s1_tbl = pd.concat([s1_tbl, temp])
#
#s1_tbl = s1_tbl.reset_index(drop=True)
#s1_tbl.columns = ['Abbrev','Long_name','Category']  
#s1_tbl.to_csv("./supp_tables/Supplementary_table_1_OptDescs.csv")
s2_tbl = pd.read_csv("./supp_tables/Supplementary_table_2_OptDescs.csv", index_col = 0)  


# Supp table 3

#odor_cid_info = pd.read_csv("./cid_info/cid_odornames.csv", index_col = 0)
#log_reg_data = pd.read_csv("./mouseOR_alignment/log_reg_auroc.csv", index_col = 0)
#log_reg_data.columns = ['PubChem CID', 'AUROC Lambda_min', 'AUROC Lambda_1se']
#
#for odor in log_reg_data['PubChem CID']:
#    o_name = odor_cid_info[odor_cid_info['cid'] == odor]['name'].values[0]
#    log_reg_data.loc[log_reg_data['PubChem CID'] == odor,'Odor'] = o_name
#    
#log_reg_data = log_reg_data.sort_values(by='AUROC Lambda_1se', ascending=False)
#log_reg_data = log_reg_data.iloc[:,[3,0,2]]
#log_reg_data.index = range(1, log_reg_data.shape[0]+1, 1)
#log_reg_data.to_csv("./supp_tables/Supplementary_table_2_LogReg.csv")

log_reg_data = pd.read_csv("./supp_tables/Supplementary_table_3_LogReg.csv", index_col = 0)

# Supp table 4

#odor_cid_info = pd.read_csv("./cid_info/cid_odornames.csv", index_col = 0)
#linear_svc = pd.read_csv("./mouseOR_alignment/linear_svc_auc.csv", index_col = 0)
#linear_svc.columns = ['PubChem CID', 'SVC AUROC']
#for odor in linear_svc['PubChem CID']:
#    o_name = odor_cid_info[odor_cid_info['cid'] == odor]['name'].values[0]
#    linear_svc.loc[linear_svc['PubChem CID'] == odor,'Odor'] = o_name
#linear_svc = linear_svc.sort_values(by='SVC AUROC', ascending=False)
#linear_svc.index = range(1, linear_svc.shape[0]+1, 1)
#linear_svc.to_csv("./supp_tables/Supplementary_table_3_LinearSVC.csv")
linear_svc = pd.read_csv("./supp_tables/Supplementary_table_4_LinearSVC.csv", index_col = 0)
